==> 00_dataimport.Rmd <==
---
title: <font color="#66C2A5">Data Import</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## Library
```{r Library}
library(Hmisc)
library(dplyr)
```


## Data
Load raw data from Clue.
```{r Data}
fulldata = read.csv(file = "ideal_partners_data.csv", sep = ",")
fulldata = fulldata[-1,-1] #first row includes labels, first column just numbers the rows
```

## Select Variables
We only need certain variables from the dataset. Columns will be renamed according to the preregistration "Political Orientation and its Relationships with Preferred Relationship Options and Partner Preferences in Single Women"
```{r Select Variables}
fulldata_selected = fulldata %>%
  select(# Political Orientation
         political_orientation = self_political_spectrum,
         
         # Age
         age = age,
         
         # Preferred Relationship Options
         interest_single = interest_being_single,
         interest_sexrel = interest_non_sexual_relations,
         interest_nonmonrel = interest_nonmonogamous_relations,
         interest_monrel = interest_monogamous_relations,
         
         # Political, Ethnical, Religious Similarity
         pref_politicalsim = lt_politcal_similarity,
         pref_ethnicalsim = lt_ethnicity_similarity,
         pref_religioussim = lt_religious_similarity,
         
         # Ideal Partner Prefs:
         pref_level_kind = lt_kindness_level,
         pref_level_supportive = lt_supportive_level,
         pref_level_attractive_body = lt_attractive_body_level,
         pref_level_attractive_face = lt_attractive_face_level,
         pref_level_financially_secure = lt_financial_security_level,
         pref_level_successful_ambitious = lt_ambition_level,
         pref_level_confident = lt_confidence_level,
         pref_level_assertive = lt_assertive_level,
         pref_level_intelligence = lt_intelligence_level,
         pref_level_educated = lt_education_level,
         
         # Ideal Age and Ideal Height:
         imp_age = lt_age_importance,
         ideal_age = lt_perfect_age,
         imp_height = lt_height_importance,
         ideal_height = lt_ideal_height,
         
         # Random Effects (Country and Language)
         country = country,
         language = language,
         
         # Reasons for exclusion
         sex = gender,
         sexual_orientation = sexual.orientation,
         relationship = recent_relationships, # the recent relationship variable needs to be wrangled later
         relationship1 = recent_relationships.1,
         relationship2 = recent_relationships.2,
         relationship3 = recent_relationships.3,
         relationship4 = recent_relationships.4,
         relationship5 = recent_relationships.5,
         answer_accuracy = answer_accuracy
         )
```

## Recode interest* variables for Danish dataset
In the Danish dataset, these variables range from 1 to 7 instead of 0 to 6
```{r}
table(fulldata_selected$interest_monrel)
fulldata_selected <- fulldata_selected %>%
  mutate(interest_single = ifelse(language == "danish", interest_single - 1, interest_single),
         interest_sexrel = ifelse(language == "danish", interest_sexrel - 1, interest_sexrel),
         interest_nonmonrel = ifelse(language == "danish", interest_nonmonrel - 1, interest_nonmonrel),
         interest_monrel = ifelse(language == "danish", interest_monrel - 1, interest_monrel))
```

## Save Dataset
```{r Save Dataset}
write.csv(fulldata_selected, file = "fulldata_selected.csv")
```
