==> 01_datawrangling.Rmd <==
---
title: <font color="#66C2A5">Data Wrangling</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## {.tabset}

### Library
```{r Library}
library(car)
library(psych)
library(ggplot2)
library(formr)
library(codebook)
library(dplyr)
```


### Data
Load selected data based on 00_dataimport
```{r Data}
fulldata_selected = read.csv(file = "fulldata_selected.csv")[,-1]
```


### Inspect Data {.tabset .active}
```{r Inspect Data}
colnames(fulldata_selected)

fulldata_selected_wrangled = fulldata_selected
```

### Age {.tabset}
Participants younger than 18 were able to insert age but the survey was ended for them afterwards.
```{r Age}
table(fulldata_selected_wrangled$age)
```
Set all numbers > 100 as missing:
```{r Age 2}
sum(fulldata_selected_wrangled$age > 100)
# This means that age is set as missing for 15 people

fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(age = ifelse(age > 100, NA, age))

table(is.na(fulldata_selected_wrangled$age))

table(fulldata_selected_wrangled$age < 18)
# 26638 people were not allowed to participate in the survey because they were younger than 18
```

### Political Orientation {.tabset}
Political Orientation item should range from 0 to 6
```{r Political Orientation}
fulldata_selected_wrangled %>%
  select("political_orientation") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$political_orientation)
```

### Preferred Relationship Options {.tabset}
Interest in being single, interest in sexual (non-romantic) relationships, interest in non-monogamous relationships and interest in monogamous relationships items should range from 0 to 6 
```{r Preferred Relationship Options}
fulldata_selected_wrangled %>%
  select(starts_with("interest_")) %>%
  lapply(table)

qplot(fulldata_selected_wrangled$interest_single)
qplot(fulldata_selected_wrangled$interest_sexrel)
qplot(fulldata_selected_wrangled$interest_nonmonrel)
qplot(fulldata_selected_wrangled$interest_monrel)
```
 

### Political, Ethnical, and Religious Similarity {.tabset}
Preferences for political, ethnical and religious similarity items should range from 0 to 6 
```{r Political, ethnical and religious similarity}
fulldata_selected_wrangled %>%
  select("pref_politicalsim", "pref_ethnicalsim", "pref_religioussim") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$pref_politicalsim)
qplot(fulldata_selected_wrangled$pref_ethnicalsim)
qplot(fulldata_selected_wrangled$pref_religioussim)
```

### Ideal Partner Preferences {.tabset}
All ideal partner preference items should range from 0 to 6 
```{r Ideal Partner Preferences}
fulldata_selected_wrangled %>%
  select(starts_with("pref_")) %>%
  select(-pref_politicalsim, -pref_ethnicalsim, -pref_religioussim) %>%
  lapply(table)

qplot(fulldata_selected_wrangled$pref_level_kind)
qplot(fulldata_selected_wrangled$pref_level_supportive)
qplot(fulldata_selected_wrangled$pref_level_attractive_body)
qplot(fulldata_selected_wrangled$pref_level_attractive_face)
qplot(fulldata_selected_wrangled$pref_level_financially_secure)
qplot(fulldata_selected_wrangled$pref_level_successful_ambitious)
qplot(fulldata_selected_wrangled$pref_level_confident)
qplot(fulldata_selected_wrangled$pref_level_assertive)
qplot(fulldata_selected_wrangled$pref_level_intelligence)
qplot(fulldata_selected_wrangled$pref_level_educated)


fulldata_selected_wrangled %>%
  select(starts_with("pref_")) %>%
  select(-pref_politicalsim, -pref_ethnicalsim, -pref_religioussim) %>%
  md_pattern()
```


### Ideal Age and Ideal Height {.tabset}
#### Importance Age
Item should range from 0 to 6 
```{r Importance Age}
fulldata_selected_wrangled %>%
  select("imp_age") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$imp_age)

```

#### Ideal Age
This item is in open format style
```{r Ideal Age}
fulldata_selected_wrangled %>%
  select("ideal_age") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$ideal_age)
```

Variable ideal_age was not restricted to a reasonable number, thus numbers larger than 100 were possible. To deal with this we will set all numbers >= 100 as missing.
```{r Ideal Age 2}
sum(fulldata_selected_wrangled$ideal_age > 100, na.rm = T)

fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(ideal_age = ifelse(ideal_age > 100, NA, ideal_age))
          
table(is.na(fulldata_selected_wrangled$ideal_age))

fulldata_selected_wrangled %>%
  select(ideal_age) %>%
  lapply(table)

qplot(fulldata_selected_wrangled$ideal_age)
```

#### Importance Height
Item should range from 0 to 6 
```{r Importance Height}
fulldata_selected_wrangled %>%
  select("imp_height") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$imp_height)
```

#### Ideal Height
Item has the options: much shorter than me, shorter than me, around the same height as me, taller than me, much taller than me
```{r Ideal Height}
table(fulldata_selected_wrangled$ideal_height)

fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(ideal_height = ifelse(ideal_height == "", NA, ideal_height),
         ideal_height_numeric = as.factor(ideal_height))

fulldata_selected_wrangled$ideal_height_numeric = 
  dplyr::recode(fulldata_selected_wrangled$ideal_height_numeric,
         'much shorter than me' = -2,'shorter than me' = -1,
         'around the same height as me' = 0, 'taller than me' = 1,
         'much taller than me' = 2, .default = NULL, .missing = NULL)

table(fulldata_selected_wrangled$ideal_height)
class(fulldata_selected_wrangled$ideal_height_numeric)
levels(fulldata_selected_wrangled$ideal_height_numeric)

fulldata_selected_wrangled %>%
  select("ideal_height_numeric") %>%
  lapply(table)

qplot(fulldata_selected_wrangled$ideal_height_numeric)

fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(ideal_height = ideal_height_numeric) %>%
  select(-ideal_height_numeric)
```


### Country {.tabset}
```{r Country}
fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(country = ifelse(country == "Korea, South",
                          "South Korea",
                          country),
         country = ifelse(country == "Congo, Democratic Republic of the",
                          "Congo, Republic of the",
                          country))

x = table(fulldata_selected_wrangled$country)
table(is.na(fulldata_selected_wrangled$country))
```


### Language {.tabset}
Chinese, Danish, English, French, German, Japanese, Portuguese, Russian, Spanish, Italian
```{r Language}
table(fulldata_selected_wrangled$language)
table(is.na(fulldata_selected_wrangled$country))
```

### Sex {.tabset}
Categories: Woman; Genderqueer/Nonbinary; Man; None of the above; Prefer not to say; Other
```{r Sex}
table(fulldata_selected_wrangled$sex)

# Set missing values as NA
fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(sex = as.character(sex),
         sex = ifelse(sex == "", NA, sex),
         sex = factor(sex))

table(fulldata_selected_wrangled$sex)

table(is.na(fulldata_selected_wrangled$sex))
```

### Sexual Orientation {.tabset}
heterosexual; Lesbian/Gay/Homosexual; Bisexual/Pansexual; Queer; Asexual; Prefer not to say; Other
```{r Sexual Orientation}
table(fulldata_selected_wrangled$sexual_orientation)

# Set missing values as NA
fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(sexual_orientation = as.character(sexual_orientation),
         sexual_orientation = ifelse(sexual_orientation == "", NA, sexual_orientation),
         sexual_orientation = factor(sexual_orientation))

table(fulldata_selected_wrangled$sexual_orientation)

table(is.na(fulldata_selected_wrangled$sexual_orientation))
```

### Relationship Variable {.tabset}
```{r Relationship variable}
table(fulldata_selected_wrangled$relationship) #No romantic or sexual relationships during the past 3 months
table(fulldata_selected_wrangled$relationship1) #Other
table(fulldata_selected_wrangled$relationship2) #Short-term (casual) sexual relationship (e.g. hookups or one-night-stands)
table(fulldata_selected_wrangled$relationship3) #New (less than 1 month old) romantic and/or sexual relationship 
table(fulldata_selected_wrangled$relationship4) #Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship
table(fulldata_selected_wrangled$relationship5)#Long-term committed/exclusive sexual relationship with one or more partners

fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(relationship = ifelse(relationship == "True",
                               "No romantic or sexual relationships during the past 3 months, ",
                               ""),
         relationship = ifelse(relationship1 == "True",
                               paste(relationship,
                                     "Other, "),
                               relationship),
         relationship = ifelse(relationship2 == "True",
                               paste(relationship,
                                     "Short-term (casual) sexual relationship (e.g. hookups or one-night-stands), "),
                               relationship),
         relationship = ifelse(relationship3 == "True",
                               paste(relationship,
                                     "New (less than 1 month old) romantic and/or sexual relationship, "),
                               relationship),
         relationship = ifelse(relationship4 == "True",
                               paste(relationship,
                                     "Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship, "),
                               relationship),
         relationship = ifelse(relationship5 == "True", 
                               paste(relationship,
                                     "Long-term committed/exclusive sexual relationship with one or more partners, "), 
                               relationship),
         relationship = ifelse(relationship == "", NA, relationship))

table(fulldata_selected_wrangled$relationship)

# Check whether command worked correctly
table(fulldata_selected_wrangled$relationship %contains% "No romantic or sexual relationships during the past 3 months") 
```

### Answer Accuracy {.tabset}
I took the survey seriously; please use my information in the study.; I did not answer seriously, please disregard my information; I choose not to answer
```{r Answer accuracy}
table(fulldata_selected_wrangled$answer_accuracy)

# Set missing values as NA
fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(answer_accuracy = as.character(answer_accuracy),
         answer_accuracy = ifelse(answer_accuracy == "", NA, answer_accuracy),
         answer_accuracy = factor(answer_accuracy))

table(fulldata_selected_wrangled$answer_accuracy)

table(is.na(fulldata_selected_wrangled$answer_accuracy))


fulldata_selected_wrangled %>%
  select(sex, answer_accuracy) %>%
  md_pattern()
```

### Creating new Variables for Ideal Partner Preferences {.tabset}
```{r}
fulldata_selected_wrangled = fulldata_selected_wrangled %>% 
  mutate(pref_level_kind_supportive = (pref_level_kind + pref_level_supportive)/2,
         pref_level_attractiveness = (pref_level_attractive_body +
                                        pref_level_attractive_face)/2,
         pref_level_financially_secure_successful_ambitious = (pref_level_financially_secure + pref_level_successful_ambitious)/2,
         pref_level_confident_assertive = (pref_level_confident + pref_level_assertive)/2, 
         pref_level_intelligence_educated = (pref_level_intelligence + pref_level_educated)/2)

table(fulldata_selected_wrangled$pref_level_kind_supportive)
table(fulldata_selected_wrangled$pref_level_attractiveness)
table(fulldata_selected_wrangled$pref_level_financially_secure_successful_ambitious)
table(fulldata_selected_wrangled$pref_level_confident_assertive)
table(fulldata_selected_wrangled$pref_level_intelligence_educated)


qplot(fulldata_selected_wrangled$pref_level_kind_supportive)
qplot(fulldata_selected_wrangled$pref_level_attractiveness)
qplot(fulldata_selected_wrangled$pref_level_financially_secure_successful_ambitious)
qplot(fulldata_selected_wrangled$pref_level_confident_assertive)
qplot(fulldata_selected_wrangled$pref_level_intelligence_educated)
```

### Save Data
```{r Save data}
write.csv(fulldata_selected_wrangled, file = "fulldata_selected_wrangled.csv")
```
