==> 02_exclusion.Rmd <==
---
title: <font color="#66C2A5">Exclusion</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---
## {.tabset}

### Library
```{r Library}
library(formr)
library(dplyr)
```

### Data
Load selected data based on 01_datawrangling
```{r Data}
fulldata_selected_wrangled = read.csv(file = "fulldata_selected_wrangled.csv")[,-1]
```

### Exclusion Process {.tabset .active}
```{r Exclusion Process}
data_included = fulldata_selected_wrangled
```

#### All participants younger than 18 {.tabset .active}
From the initial dataset of respondents (N = 94,738), `r sum(fulldata_selected_wrangled$age < 18, na.rm = T) + sum(is.na(fulldata_selected_wrangled$age))` could not participate, because they were younger than 18 years of age.
```{r Age}
data_included = fulldata_selected_wrangled %>%
  filter(age >= 18 |
         is.na(age))

nrow(data_included)

fulldata_selected_wrangled = data_included
```
The total number of responses was `r nrow(data_included)`. 

#### All participants who do not identify as women {.tabset}
From the initial dataset of respondents (n = 68,100), we excluded `r sum(fulldata_selected_wrangled$sex == "Man", fulldata_selected_wrangled$sex == "Genderqueer/Nonbinary", fulldata_selected_wrangled$sex == "None of the above", fulldata_selected_wrangled$sex == "Prefer not to say", fulldata_selected_wrangled$sex == "Other", na.rm = T)` participants because they did not identify as women (Men, Genderqueer/Nonbinary; None of the above; Prefer not to say; Other)

```{r Sex}
table(data_included$sex)

data_included = data_included %>%
  filter((sex != "Man" &
         sex != "Genderqueer/Nonbinary" &
         sex != "None of the above" &
         sex != "Prefer not to say" &
         sex != "Other")  |
         is.na(sex))

table(data_included$sex)
table(is.na(data_included$sex))
```

#### All participants who do not identify as heterosexual {.tabset}
From the initial dataset of respondents (n = 68,100), we excluded `r sum(fulldata_selected_wrangled$sexual_orientation == "Lesbian/Gay/Homosexual", fulldata_selected_wrangled$sexual_orientation == "Bisexual/Pansexual", fulldata_selected_wrangled$sexual_orientation == "Queer", fulldata_selected_wrangled$sexual_orientation == "Asexual", fulldata_selected_wrangled$sexual_orientation == "Prefer not to say", fulldata_selected_wrangled$sexual_orientation == "Other", na.rm = T)` participants because they did not identify as heterosexual (Lesbian/Gay/Homosexual; Bisexual/Pansexual; Queer; Asexual; Prefer not to say; Other)

Consecutively, this means we excluded `r sum(data_included$sexual_orientation == "Lesbian/Gay/Homosexual", data_included$sexual_orientation == "Bisexual/Pansexual", data_included$sexual_orientation == "Queer", data_included$sexual_orientation == "Asexual", data_included$sexual_orientation == "Prefer not to say", data_included$sexual_orientation == "Other", na.rm = T)` participants because they did not identify as heterosexual (Lesbian/Gay/Homosexual; Bisexual/Pansexual; Queer; Asexual; Prefer not to say; Other)  
```{r Sexuality}
table(data_included$sexual_orientation)

data_included = data_included %>%
  filter((sexual_orientation != "Lesbian/Gay/Homosexual" & sexual_orientation != "Bisexual/Pansexual" &
         sexual_orientation != "Queer" & sexual_orientation != "Asexual" &
         sexual_orientation != "Prefer not to say" & sexual_orientation != "Other") |
         is.na(sexual_orientation))

table(data_included$sexual_orientation)
table(is.na(data_included$sexual_orientation)) 
```

#### All participants who indicated being in a relationship {.tabset}
```{r}
fulldata_selected_wrangled = fulldata_selected_wrangled %>%
  mutate(relationship_binary = ifelse(is.na(relationship), 0,
                                      ifelse(relationship %contains% "New (less than 1 month old) romantic and/or sexual relationship", 1,
                                             ifelse(relationship %contains% "Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship", 1,
                                                    ifelse(relationship %contains% "Long-term committed/exclusive sexual relationship with one or more partners", 1,
                                                           ifelse(relationship %contains% "Other", 1, 0))))))
table(fulldata_selected_wrangled$relationship_binary)

data_included = data_included %>%
  mutate(relationship_binary = ifelse(is.na(relationship), 0,
                                      ifelse(relationship %contains% "New (less than 1 month old) romantic and/or sexual relationship", 1,
                                             ifelse(relationship %contains% "Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship", 1,
                                                    ifelse(relationship %contains% "Long-term committed/exclusive sexual relationship with one or more partners", 1,
                                                           ifelse(relationship %contains% "Other", 1, 0))))))
table(data_included$relationship_binary)
```

From the initial dataset of respondents (n = 68,100), we excluded `r sum(fulldata_selected_wrangled$relationship_binary, na.rm = T)` participants because they indicated being in a relationship (New (less than 1 month old) romantic and/or sexual relationship ; Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship; Long-term committed/exclusive sexual relationship with one or more partners) or participants where it was not sure whether they were currently single (Other).

Consecutively, we excluded `r sum(data_included$relationship_binary, na.rm = T)` participants because they indicated being in a relationship (New (less than 1 month old) romantic and/or sexual relationship ; Ongoing (longer than 1 month) uncommitted/non-exclusive romantic and/or sexual relationship; Long-term committed/exclusive sexual relationship with one or more partners) or participants where it was not sure whether they were currently single (Other).

```{r}
data_included = data_included %>% filter(relationship_binary == 0) %>%
  select(-relationship1, -relationship2, -relationship3, -relationship4, -relationship5, -relationship_binary)
```


#### All participants who did not specify their political orientation {.tabset}
From the initial dataset of respondents (n = 68,100), we excluded `r sum(is.na(fulldata_selected_wrangled$political_orientation))` participants because they did not specify their political orientation. 

Consecutively, we excluded `r sum(is.na(data_included$political_orientation))` participants because they did not specify their political orientation. 
```{r Political Orientation}
data_included = data_included %>%
  mutate(political_orientation = ifelse(political_orientation == "", NA, political_orientation))

table(is.na(data_included$political_orientation))

data_included = data_included %>%
  filter(!is.na(political_orientation))
```

#### All participants who did not answer the survey seriously or chose not to answer the seriousness question {.tabset}
From the initial dataset of respondents (n = 68,100), we excluded `r sum(fulldata_selected_wrangled$answer_accuracy == "I did not answer seriously; please disregard my information.", fulldata_selected_wrangled$answer_accuracy == "I choose not to answer.", na.rm = T)` participants because they indicated not answering the survey seriously (I did not answer seriously, please disregard my information; I choose not to answer)

Consecutively, we excluded `r sum(data_included$answer_accuracy == "I did not answer seriously; please disregard my information.", data_included$answer_accuracy == "I choose not to answer.", na.rm = T)` participants because they indicated not answering the survey seriously (I did not answer seriously, please disregard my information; I choose not to answer)

```{r Accuracy}
table(data_included$answer_accuracy)
data_included = data_included %>%
  filter((answer_accuracy != "I did not answer seriously; please disregard my information." &
         answer_accuracy != "I choose not to answer.") |
         is.na(answer_accuracy))
```

We have `r sum(is.na(data_included$answer_accuracy))` missings for answer_accuracy here. As mentioned in the preregistration, we included these participants in our analyses.
```{r Accuracy Missings}
missings = data_included %>% filter(is.na(answer_accuracy))
```

### Save Data
```{r Save Data}
write.csv(data_included, file = "data_included.csv")
```
