==> 03_codebook.Rmd <==
---
title: <font color="#66C2A5">Codebook</font>
csl: apa-custom-no-issue.csl
output: 
html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    code_folding: 'hide'
    self_contained: true
editor_options: 
  chunk_output_type: console
---
## {.tabset}

### Setup
Here, we're just setting a few options.
```{r Setup}
knitr::opts_chunk$set(
  warning = TRUE, # show warnings during codebook generation
  message = TRUE, # show messages during codebook generation
  error = TRUE, # do not interrupt codebook generation in case of errors,
                # usually better for debugging
  echo = TRUE  # show R code
)
ggplot2::theme_set(ggplot2::theme_bw())
pander::panderOptions("table.split.table", Inf)
```

### Library
```{r Library}
library(labelled)
library(codebook)
library(formr)
library(ggplot2)
library(dplyr)
library(future)
```

### Data
Load selected data based on 02_exclusion
```{r Data}
data_included = read.csv(file = "data_included.csv")[,-1]
```

### Information About Data {.tabset .active}
#### Variable Labels {.tabset .active}
```{r Variable Labels}
var_label(data_included) = list(
  political_orientation = "Where would you place yourself on the political spectrum?",
  age = "How old are you?",
  interest_single = "How interested are you in each of the following relationship options at some time in the future? - Being non-partnered (e.g., solo, single, not committed)",
  interest_sexrel = "How interested are you in each of the following relationship options at some time in the future? - Having sexual, non-romantic relationship(s) (e.g., hookups, one-night stands)",
  interest_nonmonrel = "How interested are you in each of the following relationship options at some time in the future? -Having non-monogamous relationship(s)",
  interest_monrel = "How interested are you in each of the following relationship options at some time in the future? - Having monogamous relationship(s)",
  pref_politicalsim = "Similar political beliefs and values as me - How important is it to you?", 
  pref_ethnicalsim = "Ethnicity/race - Should your partner's ethnicity/race be different or the same as you?",
  pref_religioussim = "Similar religious beliefs as me - How important is it to you",
  pref_level_kind = "Level of kindness - How kind should your partner be?",
  pref_level_supportive = "Supportive - How supportive should your partner be?",
  pref_level_attractive_body = "Attractive body - How attractive should your partner's body be?",
  pref_level_attractive_face = "Attractive face - How attractive should your partner's face be?",
  pref_level_financially_secure = "Financial security - How financially secure should your partner be?",
  pref_level_successful_ambitious = "Successful/ambitious - How successful/ambitious should your partner be?",
  pref_level_confident = "Level of confidence - How confident should your partner be?",
  pref_level_assertive = "Level of assertiveness - How assertive should your partner be?",
  pref_level_intelligence = "Level of intelligence - How intelligent should your partner be?",
  pref_level_educated = "Level of education - How educated should your partner be?",
  imp_age = "How important is the age of your ideal long-term partner?",
  ideal_age = "The perfect age for my ideal long-term partner is __ years old.",
  imp_height = "How important is your ideal long-term partner’s height?",
  ideal_height = "My ideal long-term partner’s height is __.",
  country = "In what country do you live?",
  language = "Language of Survey",
  sex = "Do you identify as…",
  sexual_orientation = "How would you describe your current sexual orientation?",
  relationship = "Select your relationships during the past 3 months. Choose as many as you like",
  answer_accuracy = "We understand that sometimes people fill out questionnaires for fun and give answers that may not be accurate. In the interest of scientific accuracy, we will exclude those responses from our final analysis. Please choose one of the statements below:"
)
```

#### Value Labels {.tabset}
```{r Value Labels}

val_labels(data_included$political_orientation)=
  c("LEFT" = 0, "RIGHT" = 6)

val_labels(data_included$interest_single)=
  c("not at all interested" = 0, "very interested" = 6)
val_labels(data_included$interest_sexrel)=
  c("not at all interested" = 0, "very interested" = 6)
val_labels(data_included$interest_nonmonrel)=
  c("not at all interested" = 0, "very interested" = 6)
val_labels(data_included$interest_monrel)=
  c("not at all interested" = 0, "very interested" = 6)


val_labels(data_included$pref_politicalsim)=
  c("not at all important" = 0, "very important" = 6)
val_labels(data_included$pref_ethnicalsim)=
  c("different from me" = 0, "same as me" = 6)
val_labels(data_included$pref_religioussim)=
  c("not at all important" = 0, "very important" = 6)


val_labels(data_included$pref_level_kind) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_supportive) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_attractive_body) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_attractive_face) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_financially_secure) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_successful_ambitious) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_confident) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_assertive) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_intelligence) =
  c("not at all" = 0, "very" = 6)
val_labels(data_included$pref_level_educated) =
  c("not at all" = 0, "very" = 6)

val_labels(data_included$imp_age) =
  c("not at all important" = 0, "very important" = 6)
val_labels(data_included$imp_height) =
  c("not at all important" = 0, "very important" = 6)
```

#### Relative Ideal Age {.tabset}
Ideal long-term partner's age will be be calculated as relative age discrepancy between ideal partner’s age and women’s own age which will be calculated by subtracting women’s own age from ideal partner’s age
```{r Ideal Age Range}
data_included = data_included %>%
  mutate(ideal_age_rel = ideal_age - age)
```

```{r}
qplot(data_included$ideal_age_rel)
```

### Number of participants and number of countries
```{r}
nrow(data_included)
length(unique(data_included$country))
```

### Create Codebook {.tabset}
```{r create codebook}
data_included = data_included %>%
  select(sex, country, language,
         political_orientation, age, 
          pref_politicalsim, pref_ethnicalsim, pref_religioussim,
          pref_level_financially_secure_successful_ambitious, pref_level_financially_secure, 
          pref_level_successful_ambitious,
          pref_level_confident_assertive, pref_level_confident, pref_level_assertive,
          pref_level_intelligence_educated, pref_level_intelligence, pref_level_educated,
          pref_level_kind_supportive, pref_level_kind, pref_level_supportive,
          pref_level_attractiveness, pref_level_attractive_body, pref_level_attractive_face,
          imp_age, ideal_age, ideal_age_rel,
          imp_height, ideal_height,
         interest_single, interest_sexrel, interest_nonmonrel, interest_monrel
          )

codebook(data_included)
```

### Save Data
```{r Save}
data_included_documented = data_included
write.csv(data_included_documented, file = "data_included_documented.csv")
```
