==> 05_analyses.Rmd <==
---
title: <font color="#66C2A5">Main Analyses</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## {.tabset}

### Library
```{r Library}
library(formr)
library(effects)
library(effectsize)
#library(lme4)
library(sjstats)
library(lmerTest)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(tidyverse)
library(mgcv)         
library(stringr)     
library(sandwich)     
library(lmtest) 
library(minqa)
```

### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

### Political, Ethnic, and Religious Similarity {.tabset .active}
#### H1a Preference for Similarity in Political Beliefs {.tabset}
H1a(1) There is no linear link between right-wing political orientation and women’s preferences for partner’s similar political beliefs and values.
H1a(2) There is a positive quadratic link between right-wing political orientation and women’s preferences for partner’s similar political beliefs and values.
Outcome: Preference ratings for partner's similar political beliefs and values.
Predictor: Political Orientation. Random intercept and random slope for country.


##### H1a(1) Linear Effect  {.tabset}
###### Models
```{r}
model_pref_politicalsim_lin = lmer(pref_politicalsim ~ political_orientation +
                       (1+political_orientation|country),
                               data = data_included_documented, 
                       control =lmerControl(optimizer = "bobyqa"))
```

###### Summary
```{r}
summary(model_pref_politicalsim_lin)
confint(model_pref_politicalsim_lin, level = 0.997)
```

###### Standardized Coefficients
```{r}
standardize_parameters(model_pref_politicalsim_lin, method = "refit", ci = 0.997)
```

######  Plot {.active}
```{r}
lmer(pref_politicalsim ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

##### H1a(2) Quadratic Effect {.tabset}
Here, we are examining the quadratic effect of right-wing political orientation on preferred political similarity in a partner using the Two Lines Approach (Simonsohn, 2018). We are using the Robin Hood Algorithm in order to set the breaking point. Then, we are calculating two multilevel regressions on either side of the breaking point.
Outcome: Preference ratings for partner's similar political beliefs and values.
Predictors: Political Orientation & Age. Random intercept and random slope for country.

###### Algorithm: Find breaking point
See 11_twolines_analyses_multilevel


###### Regression 1 (x <= breaking_point)
```{r}
model_pref_politicalsim_1 = lmer(pref_politicalsim ~ political_orientation +
                                 (1+political_orientation|country),
                               data = data_included_documented %>%
                                 dplyr::filter(political_orientation <= 3), control =lmerControl(optimizer = "bobyqa"))

summary(model_pref_politicalsim_1)
confint(model_pref_politicalsim_1, level = 0.997)

standardize_parameters(model_pref_politicalsim_1, method = "basic", ci = 0.997)

plot(allEffects(model_pref_politicalsim_1))
```

###### Regression 2 (x >= breaking_point)
```{r}
model_pref_politicalsim_2 = lmer(pref_politicalsim ~ political_orientation +
                                 (1+political_orientation|country),
                               data = data_included_documented %>%
                                 dplyr::filter(political_orientation >= 3), control =lmerControl(optimizer = "bobyqa"))

summary(model_pref_politicalsim_2)
confint(model_pref_politicalsim_2, level = 0.997)

standardize_parameters(model_pref_politicalsim_2, method = "basic", ci = 0.997)

plot(allEffects(model_pref_politicalsim_2))
```

#### H1b Preference for Similarity in Ethnicity/Race {.tabset}
H1b There is a positive linear link between right-wing political orientation and women’s preferences for partner’s similar ethnicity/race. 
Outcome: Preference ratings for partner's similar ethnicity/race.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_ethnicalsim <- lmer(pref_ethnicalsim ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_ethnicalsim)
confint(model_pref_ethnicalsim, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_ethnicalsim, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_ethnicalsim ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H1c Preference for Similarity in Religion {.tabset}
H1c There is a positive linear link between right-wing political orientation and women’s preferences for partner’s similar religious beliefs.  
Outcome: Preference ratings for partner's similar religious beliefs.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_religioussim <- lmer(pref_religioussim ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_religioussim)
confint(model_pref_religioussim, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_religioussim, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_religioussim ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

### Ideal Partner Preferences {.tabset}
#### H2a Preference for the Level of Financial Security- Successfulness {.tabset}
H2a There is a positive linear link between right-wing political orientation and women’s preferences for the level of financial security and successfulness.
Outcome: Level ratings for partner's financial security-successfulness.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_level_financially_secure_successful_ambitious <-
  lmer(
    pref_level_financially_secure_successful_ambitious ~ political_orientation + (1 +
                                                                                    political_orientation |
                                                                                    country),
    data = data_included_documented,
    control = lmerControl(optimizer = "bobyqa")
  )
```

##### Summary
```{r}
summary(model_pref_level_financially_secure_successful_ambitious)
confint(model_pref_level_financially_secure_successful_ambitious, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_level_financially_secure_successful_ambitious, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_level_financially_secure_successful_ambitious ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H2b Preference for the Level of Confidence-Assertiveness {.tabset}
H2b There is a positive linear link between right-wing political orientation and women’s preferences for the level of confidence and assertiveness.  
Outcome: Level ratings for partner's confidence-assertiveness.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_level_confident_assertive <- lmer(pref_level_confident_assertive ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_level_confident_assertive)
confint(model_pref_level_confident_assertive, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_level_confident_assertive, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_level_confident_assertive ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```


#### H2c Preference for the Level of Education-Intelligence {.tabset}
H2c There is no link between right-wing political orientation and women’s preferences for the level of education and intelligence.  
Outcome: Level ratings for partner's education-intelligence.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_level_intelligence_educated <- lmer(pref_level_intelligence_educated ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_level_intelligence_educated)
confint(model_pref_level_intelligence_educated, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_level_intelligence_educated, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_level_intelligence_educated ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```


#### H2d Preference for the Level of Kindness-Supportiveness {.tabset}
H2d There is no link between right-wing political orientation and women’s preferences for the level of kindness and supportiveness.  
Outcome: Level ratings for partner's kindness-supportiveness.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_level_kind_supportive <- lmer(pref_level_kind_supportive ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_level_kind_supportive)
confint(model_pref_level_kind_supportive, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_level_kind_supportive, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_level_kind_supportive ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H2e Preference for the Level of Attractiveness {.tabset}
H2e There is no link between right-wing political orientation and women’s preferences for the level of attractiveness.  
Outcome: Level ratings for partner's attractiveness.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_pref_level_attractiveness <- lmer(pref_level_attractiveness ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_pref_level_attractiveness)
confint(model_pref_level_attractiveness, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_pref_level_attractiveness, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(pref_level_attractiveness ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```



### Ideal Age and Height {.tabset}
#### H3a(1) Importance Ratings for Partner's Age {.tabset}
H3a(1) There is a positive linear link between right-wing political orientation and women’s importance ratings for partner’s age. 
Outcome: Importance ratings for partner's age.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_imp_age <- lmer(imp_age ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_imp_age)
confint(model_imp_age, level= 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_imp_age, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(imp_age ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H3a(2) Level Ratings for Partner's Age {.tabset}
H3a(2) There is a positive linear link between right-wing political orientation and the relative age discrepancy between ideal partner’s age and women’s age. (discrepancy calculated as ideal partner’s age – women’s age)
Outcome: Discrepancy between level ratings for ideal partner's age and women's age (ideal_age_rel)
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_ideal_age_rel <- lmer(ideal_age_rel ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_ideal_age_rel)
confint(model_ideal_age_rel, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_ideal_age_rel, method = "basic", ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(ideal_age_rel ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H3b(1) Importance Ratings for Partner's Height {.tabset}
H3b(1) There is a positive linear link between right-wing political orientation and women’s importance ratings for partner’s height.
Outcome: Importance ratings for partner's height.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_imp_height <- lmer(imp_height ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_imp_height)
confint(model_imp_height, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_imp_height, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(imp_height ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### H3b(2) Level Ratings for Partner's Height {.tabset}
H3b(2) There is a positive linear link between right-wing political orientation and ideal partner’s height.
Outcome: Level ratings for ideal partner's height.
Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_ideal_height <- lmer(ideal_height ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_ideal_height)
confint(model_ideal_height, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_ideal_height, method = "basic", ci = 0.997)
```

#####  Plot {.active}
```{r}
lmer(ideal_height ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```
