==> 07_graphs.Rmd <==
---
title: <font color="#66C2A5">Graphs</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## {.tabset}

### Library
```{r Library}
library(formr)
library(effects)
library(effectsize)
library(lme4)
library(sjstats)
library(lmerTest)
library(ggplot2)
library(tidyr)
library(ggpubr)
library(RColorBrewer)
library(coefplot)
library(dplyr)
library(purrr) # for running multiple regression
library(broom)
library(lm.beta)
library(stringr)
library(tidyr)
library(tidyverse)
library(countrycode)


apatheme = theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    panel.border = element_blank(),
    axis.line = element_line(),
    legend.title = element_blank(),
    plot.title = element_text(hjust = 0.5)
  )
```

### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[, -1]
data_included_documented = data_included_documented %>%
  filter(!is.na(political_orientation)) #check
```


### Figure 1{.tabset}
```{r Political Orientation}
mean(data_included_documented$political_orientation) %>% round(., 2)
sd(data_included_documented$political_orientation) %>% round(., 2)
range(data_included_documented$political_orientation)

hist <- ggplot(data_included_documented, aes(x = political_orientation)) + 
  geom_histogram(col = "grey", binwidth = 0.5, center = 0) +
  labs(x = "Political Orientation", y = "Number of Participants")+ 
  theme(text = element_text(size=15), axis.text.x = element_text(size = 10),
        axis.text.y = element_text(size = 10))+
 scale_x_continuous(breaks=c(0, 1, 2, 3, 4, 5, 6))+
  apatheme

hist
```


#### Save Image
```{r}
jpeg("kuschel-figure1.jpeg", width = 2*1580, height = 2*836, res = 600)
hist
dev.off()
```


### Figure 3 {.tabset}
Effect size estimates of linear effects of political orientation on preferred political, ethnical and religious similarity for main analyses and one robustness analysis. Graph further includes both linear effect size estimates according to the Two Lines Approach. 

#### Main Analyses
```{r}
# H1a(1) Linear Effect on Political Similarity
model_pref_politicalsim_lin <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~
        political_orientation +
        (1 + political_orientation |
           country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

#H1a(2) Quadratic Effect on Political Similarity
model_pref_politicalsim_1 <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~ political_orientation +                       (1 + political_orientation |
                                                                           country),
      data = data_included_documented
      %>%                     dplyr::filter(political_orientation <= 3),
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_politicalsim_2 <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~ political_orientation +                       (1 + political_orientation |
                                                                           country),
      data = data_included_documented
      %>%                     dplyr::filter(political_orientation >= 3),
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_ethnicalsim <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_ethnicalsim ~ political_orientation + (1 + political_orientation |
                                                    country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_religioussim <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_religioussim ~ political_orientation + (1 + political_orientation |
                                                     country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Robustness Analyses
Additionally including age as a predictor
```{r}
# H1a(1) Linear Effect on Political Similarity
model_pref_politicalsim_lin_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~ political_orientation  + age +
        (1 + political_orientation | country),
      data =
        data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

#H1a(2) Quadratic Effect on Political Similarity

model_pref_politicalsim_1_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~ political_orientation + age +
        (1 + political_orientation |
           country),
      data = data_included_documented %>%
        dplyr::filter(political_orientation <= 3),
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_politicalsim_2_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_politicalsim ~ political_orientation + age +
        (1 + political_orientation |
           country),
      data = data_included_documented %>%
        dplyr::filter(political_orientation >= 3),
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_ethnicalsim_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_ethnicalsim ~ political_orientation + age + (1 + political_orientation |
                                                          country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_religioussim_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_religioussim ~ political_orientation + age + (1 + political_orientation |
                                                           country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```


#### Combine {.active}
```{r}
all_models_similarity <- rbind(
  model_pref_politicalsim_lin,
  model_pref_politicalsim_1,
  model_pref_politicalsim_2,
  model_pref_ethnicalsim,
  model_pref_religioussim,
  model_pref_politicalsim_lin_robustcheck,
  model_pref_politicalsim_1_robustcheck,
  model_pref_politicalsim_2_robustcheck,
  model_pref_ethnicalsim_robustcheck,
  model_pref_religioussim_robustcheck
) %>%
  filter(Parameter == "political_orientation")



all_models_similarity = rbind(
  all_models_similarity[1, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[2:3, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[4, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[5:6, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[7:8, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[9, ],
  c("political_orientation", NA, 0.997, NA, NA),
  all_models_similarity[10, ]
)


all_models_similarity$hypothesis = factor(
  c(
    "H1a(1) Preferred \nPolitical Similarity \n(Linear Effect)",
    "",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO <= 3",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO >= 3",
    "",
    "H1b) Preferred \nEthnic Similarity",
    "   ",
    "H1c) Preferred \nReligious Similarity",
    "H1a(1) Preferred \nPolitical Similarity \n(Linear Effect)",
    " ",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO <= 3",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO >= 3",
    " ",
    "H1b) Preferred \nEthnic Similarity",
    "   ",
    "H1c) Preferred \nReligious Similarity"
  ),
  levels = c(
    "H1c) Preferred \nReligious Similarity",
    "   ",
    "H1b) Preferred \nEthnic Similarity",
    " ",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO >= 3",
    "H1a(2) Preferred \nPolitical Similarity \n(Quadratic Effect) \nPO <= 3",
    "",
    "H1a(1) Preferred \nPolitical Similarity \n(Linear Effect)"
  )
)

all_models_similarity$type = factor(c(rep("Main Analyses", 8), rep("Robustness Analyses", 8)),
                                    levels = c("Robustness Analyses", "Main Analyses"))

all_models_similarity$Std_Coefficient = as.numeric(all_models_similarity$Std_Coefficient)

all_models_similarity$CI_low = as.numeric(all_models_similarity$CI_low)

all_models_similarity$CI_high = as.numeric(all_models_similarity$CI_high)

multiplot = ggplot(all_models_similarity, aes(hypothesis, Std_Coefficient)) +
  geom_pointrange(
    data = all_models_similarity,
    mapping = aes(
      x = hypothesis,
      y = Std_Coefficient,
      ymin = CI_low,
      ymax = CI_high,
      group = type,
      color = type
    ),
    position = position_dodge(width = c(0.3))
  ) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  coord_flip() +
  apatheme +
  scale_colour_manual(values = c("#86BDDB", "#084594"))  +
  labs(y = "Linear Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 15),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 10),
    legend.text = element_text(size = 10),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE)) +
  theme(axis.ticks.y = element_blank())
multiplot
```


#### Save image
```{r}
jpeg("kuschel-figure3.jpeg",
     width = 6300,
     height = 4200,
     res = 600)
multiplot
dev.off()
```


### Figure 4 {.tabset}
Effect size estimates of quadratic effect of political orientation on preferred political similarity for main analysis and one robustness analysis.  

#### Analyses {.active}
```{r}
data_included_nomiss_politicalsim_smaller = data_included_documented %>%
  filter(!is.na(pref_politicalsim)) %>%
  filter(political_orientation <= 3)


model_pref_politicalsim_1 = lmer(
  pref_politicalsim ~ political_orientation +
    (1 + political_orientation |
       country),
  data = data_included_nomiss_politicalsim_smaller,
  control = lmerControl(optimizer = "bobyqa")
)
pred_pref_politicalsim_1 = predict(model_pref_politicalsim_1)

data_included_nomiss_politicalsim_smaller$pred_pref_politicalsim_1 =
  pred_pref_politicalsim_1


data_included_nomiss_politicalsim_larger = data_included_documented %>%
  filter(!is.na(pref_politicalsim)) %>%
  filter(political_orientation >= 3)


model_pref_politicalsim_1 = lmer(
  pref_politicalsim ~ political_orientation +
    (1 + political_orientation |
       country),
  data = data_included_nomiss_politicalsim_larger,
  control = lmerControl(optimizer = "bobyqa")
)
pred_pref_politicalsim_1 = predict(model_pref_politicalsim_1)

data_included_nomiss_politicalsim_larger$pred_pref_politicalsim_1 =
  pred_pref_politicalsim_1

data_included_nomiss_politicalsim = rbind(
  data_included_nomiss_politicalsim_smaller,
  data_included_nomiss_politicalsim_larger
)

#Plotting empirical values
graph_emp =
  ggplot(data = data_included_nomiss_politicalsim,
         aes(x = political_orientation, y = pred_pref_politicalsim_1)) +
  geom_jitter(aes(y = pref_politicalsim),
              alpha = 0.5,
              color = "grey44") +
  geom_smooth(
    method = 'lm',
    formula = y ~ x,
    fill = "#B22222",
    color = "#B22222",
    level = 0.997,
    data = data_included_nomiss_politicalsim %>%
      filter(political_orientation <= 3)
  ) +
  geom_smooth(
    method = 'lm',
    formula = y ~ x,
    fill = "#084594",
    color = "#084594",
    level = 0.997,
    data = data_included_nomiss_politicalsim %>%
      filter(political_orientation >= 3)
  ) +
  geom_vline(xintercept = 3,
             linetype = "dashed",
             linewidth = 1) +
  apatheme +
  labs(y = "Preferred Political Similarity\n(Empirical Values)", x = "Political Orientation") +
  scale_x_continuous(breaks = c(0, 1, 2, 3, 4, 5, 6)) +
  scale_y_continuous(breaks = c(0, 1, 2, 3, 4, 5, 6)) +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15)
  )

graph_emp
```


#### Save image
```{r}
jpeg("kuschel-figure4.jpeg",
     width = 6000,
     height = 6000,
     res = 600)
graph_emp
dev.off()
```

### Figure 5 {.tabset .active}
Effect size estimates of linear effects of political orientation on ideal partner preferences for main analyses and one robustness analysis.

#### Main Analyses
```{r}
model_pref_level_financially_secure_successful_ambitious <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_financially_secure_successful_ambitious ~ political_orientation + (1 +
                                                                                      political_orientation | country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_confident_assertive <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_confident_assertive ~ political_orientation + (1 + political_orientation |
                                                                  country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_intelligence_educated <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_intelligence_educated ~ political_orientation + (1 + political_orientation |
                                                                    country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_kind_supportive <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_kind_supportive ~ political_orientation + (1 + political_orientation |
                                                              country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_attractiveness <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_attractiveness ~ political_orientation + (1 + political_orientation |
                                                             country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Robustness Analyses
Additionally including age as a predictor
```{r}
model_pref_level_financially_secure_successful_ambitious_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_financially_secure_successful_ambitious ~ political_orientation + age + (1 +
                                                                                            political_orientation | country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_confident_assertive_robustcheck  <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_confident_assertive ~ political_orientation + age + (1 + political_orientation |
                                                                        country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_intelligence_educated_robustcheck  <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_intelligence_educated ~ political_orientation + age + (1 + political_orientation |
                                                                          country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_kind_supportive_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_kind_supportive ~ political_orientation + age + (1 + political_orientation |
                                                                    country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_pref_level_attractiveness_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      pref_level_attractiveness ~ political_orientation + age + (1 + political_orientation |
                                                                   country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Combine {.active}
```{r}
all_models_ipp <- rbind(
  model_pref_level_financially_secure_successful_ambitious,
  model_pref_level_confident_assertive,
  model_pref_level_intelligence_educated,
  model_pref_level_kind_supportive,
  model_pref_level_attractiveness,
  model_pref_level_financially_secure_successful_ambitious_robustcheck,
  model_pref_level_confident_assertive_robustcheck,
  model_pref_level_intelligence_educated_robustcheck,
  model_pref_level_kind_supportive_robustcheck,
  model_pref_level_attractiveness_robustcheck
) %>%
  filter(Parameter == "political_orientation")

all_models_ipp$hypothesis = factor(
  rep(
    c(
      "H2a) Financial Security-\nSuccessfulness",
      "H2b) Confidence-\nAssertiveness",
      "H2c) Education-\nIntelligence",
      "H2d) Kindness-\nSupportiveness",
      "H2e) Attractiveness"
    ),
    2
  ),
  levels = c(
    "H2e) Attractiveness",
    "H2d) Kindness-\nSupportiveness",
    "H2c) Education-\nIntelligence",
    "H2b) Confidence-\nAssertiveness",
    "H2a) Financial Security-\nSuccessfulness"
  )
)

all_models_ipp$type = factor(c(rep("Main Analyses", 5), rep("Robustness Analyses", 5)),
                             levels = c("Robustness Analyses", "Main Analyses"))


multiplot = ggplot(all_models_ipp, aes(hypothesis, Std_Coefficient)) +
  geom_pointrange(
    data = all_models_ipp,
    mapping = aes(
      x = hypothesis,
      y = Std_Coefficient,
      ymin = CI_low,
      ymax = CI_high,
      group = type,
      color = type
    ),
    position = position_dodge(width = 0.3)
  ) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  coord_flip() +
  apatheme +
  scale_colour_manual(values = c("#86BDDB", "#084594"))  +
  labs(y = "Linear Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 15),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 10),
    legend.text = element_text(size = 10),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) + guides(col = guide_legend(reverse = TRUE))
multiplot
```

#### Save image
```{r}
jpeg("kuschel-figure5.jpeg",
     width = 6300,
     height = 4200,
     res = 600)
multiplot
dev.off()
```

### Figure 76{.tabset}
Effect size estimates of linear effects of political orientation on ideal partner's age and height for main analyses and one robustness analysis.

#### Main Analyses
```{r}
model_imp_age <-
  as.data.frame(standardize_parameters(
    lmer(
      imp_age ~ political_orientation + (1 + political_orientation |
                                           country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_ideal_age_rel <-
  as.data.frame(standardize_parameters(
    lmer(
      ideal_age_rel ~ political_orientation + (1 + political_orientation |
                                                 country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_imp_height <-
  as.data.frame(standardize_parameters(
    lmer(
      imp_height ~ political_orientation + (1 + political_orientation |
                                              country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_ideal_height <-
  as.data.frame(standardize_parameters(
    lmer(
      ideal_height ~ political_orientation + (1 + political_orientation |
                                                country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Robustness Analyses
Additionally including age as a predictor
```{r}
model_imp_age_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      imp_age ~ political_orientation + age + (1 + political_orientation |
                                                 country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_ideal_age_rel_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      ideal_age_rel ~ political_orientation + age + (1 + political_orientation |
                                                       country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_imp_height_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      imp_height ~ political_orientation + age + (1 + political_orientation |
                                                    country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_ideal_height_robustcheck  <-
  as.data.frame(standardize_parameters(
    lmer(
      ideal_height ~ political_orientation + age + (1 + political_orientation |
                                                      country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Combine {.active}
```{r}
all_models_ageheight <- rbind(
  model_imp_age,
  model_ideal_age_rel,
  model_imp_height,
  model_ideal_height,
  model_imp_age_robustcheck,
  model_ideal_age_rel_robustcheck,
  model_imp_height_robustcheck,
  model_ideal_height_robustcheck
) %>%
  filter(Parameter == "political_orientation")

all_models_ageheight$hypothesis = factor(
  rep(
    c(
      "H3a(1)) Ideal Age \n(Importance)",
      "H3a(2)) Ideal Age \n(Level)",
      "H3b(1)) Ideal Height \n(Importance)",
      "H3b(2)) Ideal Height \n(Level)"
    ),
    2
  ),
  levels = c(
    "H3b(2)) Ideal Height \n(Level)",
    "H3b(1)) Ideal Height \n(Importance)",
    "H3a(2)) Ideal Age \n(Level)",
    "H3a(1)) Ideal Age \n(Importance)"
  )
)

all_models_ageheight$type = factor(c(rep("Main Analyses", 4), rep("Robustness Analyses", 4)),
                                   levels = c("Robustness Analyses", "Main Analyses"))


multiplot = ggplot(all_models_ageheight, aes(hypothesis, Std_Coefficient)) +
  geom_pointrange(
    data = all_models_ageheight,
    mapping = aes(
      x = hypothesis,
      y = Std_Coefficient,
      ymin = CI_low,
      ymax = CI_high,
      group = type,
      color = type
    ),
    position = position_dodge(width = 0.3)
  ) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  coord_flip() +
  apatheme +
  scale_colour_manual(values = c("#86BDDB", "#084594"))  +
  labs(y = "Linear Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 15),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 10),
    legend.text = element_text(size = 10),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) + guides(col = guide_legend(reverse = TRUE))
multiplot
```

#### Save image
```{r}
jpeg("kuschel-figure6.jpeg",
     width = 6300,
     height = 4200,
     res = 600)
multiplot
dev.off()
```


### Figure S2



#### Inclusion of data
```{r}
countries = as.data.frame(table(data_included_documented$country)) %>%
  arrange(-Freq)

countries
```
We will include all countries with more than 500 participants. This allows us to show effect sizes for a diverse range of countries.
Diversity of countries is indicated by:

* location: European (France, Germany, Italy, Spain); North American (United States of America); South American (Mexico, Brazil)
* language: French (France); German (Germany); English (United States of America); Spanish (Mexico, Spain); Italian (Italy); Portuguese (Brazil)
* culture: Western (France, Germany, Italy, Spain, United States of America); Non-Western (Mexico, Brazil)

Sample sizes of other countries are too small (n < 500) to reach any conclusions.

```{r}
seven_countries = countries %>% filter(Freq > 500)
data_included_documented_rescon = data_included_documented %>%
  filter(country %in% seven_countries$Var1)
```

```{r}
countries_rescon = as.data.frame(table(data_included_documented_rescon$country)) %>%
  arrange(-Freq)
countries_rescon
```

#### Models
##### H1a Preference for similarity in political beliefs
###### H1a(1) Linear Effect
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_politicalsim,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_politicalsim = data_included_documented_rescon_wide %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(data_included_documented_rescon_wide$pref_politicalsim) ~ scale(.x),
    data = data_included_documented_rescon_wide
  )) %>%
  map(lm.beta)

models_pref_politicalsim_lin_coef = models_pref_politicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_lin_confint_low = models_pref_politicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_lin_confint_high = models_pref_politicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_lin_graph = left_join(
  left_join(
    models_pref_politicalsim_lin_coef,
    models_pref_politicalsim_lin_confint_low,
    by = "name"
  ),
  models_pref_politicalsim_lin_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1a(1)) Preferred Political \nSimilarity - Linear Effect")
```

###### H1a(2) Quadratic Effect
####### Regression 1 (x <= breaking_point)
```{r}
data_included_documented_rescon_wide_reg1 = data_included_documented_rescon %>%
  dplyr::filter(political_orientation <= 3) %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_politicalsim,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )



models_pref_politicalsim_reg1 = data_included_documented_rescon_wide_reg1 %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(
      data_included_documented_rescon_wide_reg1$pref_politicalsim
    ) ~
      scale(.x),
    data = data_included_documented_rescon_wide_reg1
  )) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg1 = models_pref_politicalsim_reg1 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_quad_confint_low_reg1 = models_pref_politicalsim_reg1 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_quad_confint_high_reg1 = models_pref_politicalsim_reg1 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_quad_graph_reg1 = left_join(
  left_join(
    models_pref_politicalsim_quad_coef_reg1,
    models_pref_politicalsim_quad_confint_low_reg1,
    by = "name"
  ),
  models_pref_politicalsim_quad_confint_high_reg1,
  by = "name"
) %>%
  mutate(outcome = "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO <= 3")
```


####### Regression 2 (x >= breaking_point)
```{r}
data_included_documented_rescon_wide_reg2 = data_included_documented_rescon %>%
  dplyr::filter(political_orientation >= 3) %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_politicalsim,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_politicalsim_reg2 = data_included_documented_rescon_wide_reg2 %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(
      data_included_documented_rescon_wide_reg2$pref_politicalsim
    ) ~
      scale(.x),
    data = data_included_documented_rescon_wide_reg2
  )) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg2 = models_pref_politicalsim_reg2 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_quad_confint_low_reg2 = models_pref_politicalsim_reg2 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_quad_confint_high_reg2 = models_pref_politicalsim_reg2 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_quad_graph_reg2 = left_join(
  left_join(
    models_pref_politicalsim_quad_coef_reg2,
    models_pref_politicalsim_quad_confint_low_reg2,
    by = "name"
  ),
  models_pref_politicalsim_quad_confint_high_reg2,
  by = "name"
) %>%
  mutate(outcome = "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO >= 3")
```

##### H1b Preference for similarity in ethnicity/race
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_ethnicalsim,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_ethnicalsim = data_included_documented_rescon_wide %>%
  select(-pref_ethnicalsim) %>%
  map( ~ lm(
    scale(data_included_documented_rescon_wide$pref_ethnicalsim) ~ scale(.x) +
      I(scale(.x) ^ 2),
    data = data_included_documented_rescon_wide
  )) %>%
  map(lm.beta)

models_pref_ethnicalsim_coef = models_pref_ethnicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_ethnicalsim_confint_low = models_pref_ethnicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_ethnicalsim_confint_high = models_pref_ethnicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_ethnicalsim_graph = left_join(
  left_join(
    models_pref_ethnicalsim_coef,
    models_pref_ethnicalsim_confint_low,
    by = "name"
  ),
  models_pref_ethnicalsim_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1b) Preferred \nEthnic Similarity")
```

##### H1c Preference for similarity in religion
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_religioussim,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_religioussim = data_included_documented_rescon_wide %>%
  select(-pref_religioussim) %>%
  map( ~ lm(
    scale(data_included_documented_rescon_wide$pref_religioussim) ~ scale(.x) +
      I(scale(.x) ^ 2),
    data = data_included_documented_rescon_wide
  )) %>%
  map(lm.beta)

models_pref_religioussim_coef = models_pref_religioussim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_religioussim_confint_low = models_pref_religioussim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_religioussim_confint_high = models_pref_religioussim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_religioussim_graph = left_join(
  left_join(
    models_pref_religioussim_coef,
    models_pref_religioussim_confint_low,
    by = "name"
  ),
  models_pref_religioussim_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1c) Preferred \nReligious Similarity")
```

##### H2a Preference for the level of financial security- successfulness 
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_level_financially_secure_successful_ambitious,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_level_financially_secure_successful_ambitious = data_included_documented_rescon_wide %>%
  select(-pref_level_financially_secure_successful_ambitious) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$pref_level_financially_secure_successful_ambitious ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_financially_secure_successful_ambitious_coef = models_pref_level_financially_secure_successful_ambitious %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_financially_secure_successful_ambitious_confint_low = models_pref_level_financially_secure_successful_ambitious %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_financially_secure_successful_ambitious_confint_high = models_pref_level_financially_secure_successful_ambitious %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_financially_secure_successful_ambitious_graph = left_join(
  left_join(
    models_pref_level_financially_secure_successful_ambitious_coef,
    models_pref_level_financially_secure_successful_ambitious_confint_low,
    by = "name"
  ),
  models_pref_level_financially_secure_successful_ambitious_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2a) Financial Security-\nSuccessfulness")
```

##### H2b Preference for the level of confidence-assertiveness
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_level_confident_assertive,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_level_confident_assertive = data_included_documented_rescon_wide %>%
  select(-pref_level_confident_assertive) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$pref_level_confident_assertive ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_confident_assertive_coef = models_pref_level_confident_assertive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_confident_assertive_confint_low = models_pref_level_confident_assertive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_confident_assertive_confint_high = models_pref_level_confident_assertive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_confident_assertive_graph = left_join(
  left_join(
    models_pref_level_confident_assertive_coef,
    models_pref_level_confident_assertive_confint_low,
    by = "name"
  ),
  models_pref_level_confident_assertive_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2b) Confidence-\nAssertiveness")
```

##### H2c Preference for the level of education-intelligence
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_level_intelligence_educated,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_level_intelligence_educated = data_included_documented_rescon_wide %>%
  select(-pref_level_intelligence_educated) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$pref_level_intelligence_educated ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_intelligence_educated_coef = models_pref_level_intelligence_educated %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_intelligence_educated_confint_low = models_pref_level_intelligence_educated %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_intelligence_educated_confint_high = models_pref_level_intelligence_educated %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_intelligence_educated_graph = left_join(
  left_join(
    models_pref_level_intelligence_educated_coef,
    models_pref_level_intelligence_educated_confint_low,
    by = "name"
  ),
  models_pref_level_intelligence_educated_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2c) Education-\nIntelligence")
```

##### H2d Preference for the level of kindness-supportiveness
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_level_kind_supportive,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_level_kind_supportive = data_included_documented_rescon_wide %>%
  select(-pref_level_kind_supportive) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$pref_level_kind_supportive ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_kind_supportive_coef = models_pref_level_kind_supportive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_kind_supportive_confint_low = models_pref_level_kind_supportive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_kind_supportive_confint_high = models_pref_level_kind_supportive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_kind_supportive_graph = left_join(
  left_join(
    models_pref_level_kind_supportive_coef,
    models_pref_level_kind_supportive_confint_low,
    by = "name"
  ),
  models_pref_level_kind_supportive_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2d) Kindness-\nSupportiveness")
```

##### H2e Preference for the level of attractiveness
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(
    pref_level_attractiveness,
    France,
    Germany,
    `United States of America`,
    Mexico,
    Italy,
    Brazil,
    Spain
  )

models_pref_level_attractiveness = data_included_documented_rescon_wide %>%
  select(-pref_level_attractiveness) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$pref_level_attractiveness ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_attractiveness_coef = models_pref_level_attractiveness %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_attractiveness_confint_low = models_pref_level_attractiveness %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_attractiveness_confint_high = models_pref_level_attractiveness %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_attractiveness_graph = left_join(
  left_join(
    models_pref_level_attractiveness_coef,
    models_pref_level_attractiveness_confint_low,
    by = "name"
  ),
  models_pref_level_attractiveness_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2e) Attractiveness")
```


##### H3a(1) Ideal age (importance)
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(imp_age,
         France,
         Germany,
         `United States of America`,
         Mexico,
         Italy,
         Brazil,
         Spain)

models_imp_age = data_included_documented_rescon_wide %>%
  select(-imp_age) %>%
  map(
    ~ lm(data_included_documented_rescon_wide$imp_age ~ .x,
         data = data_included_documented_rescon_wide)
  ) %>%
  map(lm.beta)

models_imp_age_coef = models_imp_age %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_age_confint_low = models_imp_age %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_imp_age_confint_high = models_imp_age %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_imp_age_graph = left_join(
  left_join(models_imp_age_coef,
            models_imp_age_confint_low,
            by = "name"),
  models_imp_age_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3a(1)) Ideal Age \n(Importance)")
```

##### H3a(2) Ideal age (level)
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(ideal_age_rel,
         France,
         Germany,
         `United States of America`,
         Mexico,
         Italy,
         Brazil,
         Spain)

models_ideal_age_rel = data_included_documented_rescon_wide %>%
  select(-ideal_age_rel) %>%
  map(
    ~ lm(
      data_included_documented_rescon_wide$ideal_age_rel ~ .x,
      data = data_included_documented_rescon_wide
    )
  ) %>%
  map(lm.beta)

models_ideal_age_rel_coef = models_ideal_age_rel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_age_rel_confint_low = models_ideal_age_rel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_ideal_age_rel_confint_high = models_ideal_age_rel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_ideal_age_rel_graph = left_join(
  left_join(
    models_ideal_age_rel_coef,
    models_ideal_age_rel_confint_low,
    by = "name"
  ),
  models_ideal_age_rel_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3a(2)) Ideal Age \n(Level)")
```

##### H3b(1) Ideal height (importance)
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(imp_height,
         France,
         Germany,
         `United States of America`,
         Mexico,
         Italy,
         Brazil,
         Spain)

models_imp_height = data_included_documented_rescon_wide %>%
  select(-imp_height) %>%
  map(
    ~ lm(data_included_documented_rescon_wide$imp_height ~ .x,
         data = data_included_documented_rescon_wide)
  ) %>%
  map(lm.beta)

models_imp_height_coef = models_imp_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_height_confint_low = models_imp_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_imp_height_confint_high = models_imp_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_imp_height_graph = left_join(
  left_join(models_imp_height_coef,
            models_imp_height_confint_low,
            by = "name"),
  models_imp_height_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3b(1)) Ideal Height \n(Importance)")
```

##### H3b(2) Ideal height (level)
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(ideal_height,
         France,
         Germany,
         `United States of America`,
         Mexico,
         Italy,
         Brazil,
         Spain)

models_ideal_height = data_included_documented_rescon_wide %>%
  select(-ideal_height) %>%
  map(
    ~ lm(data_included_documented_rescon_wide$ideal_height ~ .x,
         data = data_included_documented_rescon_wide)
  ) %>%
  map(lm.beta)

models_ideal_height_coef = models_ideal_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_height_confint_low = models_ideal_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_ideal_height_confint_high = models_ideal_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_ideal_height_graph = left_join(
  left_join(
    models_ideal_height_coef,
    models_ideal_height_confint_low,
    by = "name"
  ),
  models_ideal_height_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3b(2)) Ideal Height \n(Level)")
```

#### Combine Graph One
```{r}
data_graph = rbind(
  models_pref_politicalsim_lin_graph,
  models_pref_politicalsim_quad_graph_reg1,
  models_pref_politicalsim_quad_graph_reg2,
  models_pref_ethnicalsim_graph,
  models_pref_religioussim_graph
) %>%
  mutate(outcome = factor(
    outcome,
    levels = c(
      "H1c) Preferred \nReligious Similarity",
      "H1b) Preferred \nEthnic Similarity",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO >= 3",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO <= 3",
      "H1a(1)) Preferred Political \nSimilarity - Linear Effect"
    )
  ))


data_graph$name = factor(rep(
  c(
    paste0(
      "France (n = ",
      sum(data_included_documented_rescon$country == "France"),
      ")"
    ),
    paste0(
      "Germany (n = ",
      sum(data_included_documented_rescon$country == "Germany"),
      ")"
    ),
    paste0(
      "United States of America (n = ",
      sum(
        data_included_documented_rescon$country == "United States of America"
      ),
      ")"
    ),
    paste0(
      "Mexico (n = ",
      sum(data_included_documented_rescon$country == "Mexico"),
      ")"
    ),
    paste0(
      "Italy (n = ",
      sum(data_included_documented_rescon$country == "Italy"),
      ")"
    ),
    paste0(
      "Brazil (n = ",
      sum(data_included_documented_rescon$country == "Brazil"),
      ")"
    ),
    paste0(
      "Spain (n = ",
      sum(data_included_documented_rescon$country == "Spain"),
      ")"
    )
  ),
  5
))

l_countries = levels(data_graph$name)
data_graph$name = factor(data_graph$name,
                         levels = rev(
                           c(
                             l_countries[2],
                             l_countries[3],
                             l_countries[7],
                             l_countries[5],
                             l_countries[4],
                             l_countries[1],
                             l_countries[6]
                           )
                         ))
```

#### Graph One Multiplot {.active}
```{r}
multiplot = ggplot(data_graph, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_pointrange(
    data = data_graph,
    mapping = aes(
      x = outcome,
      y = mean,
      ymin = conf_low ,
      ymax = conf_high ,
      group = name,
      color = name
    ),
    position = position_dodge(width = 0.4)
  ) +
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 15),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE))
multiplot
```

##### Save image
```{r}
jpeg(
  "kuschel-figureS2_1.jpeg",
  width = 2*5000,
  height = 2*3000,
  res = 600
)
multiplot
dev.off()
```

#### Combine Graph Two
```{r}
data_graph2 = rbind(
  models_pref_level_financially_secure_successful_ambitious_graph,
  models_pref_level_kind_supportive_graph,
  models_pref_level_attractiveness_graph,
  models_pref_level_confident_assertive_graph,
  models_pref_level_intelligence_educated_graph,
  models_imp_age_graph,
  models_ideal_age_rel_graph,
  models_imp_height_graph,
  models_ideal_height_graph
) %>%
  mutate(outcome = factor(
    outcome,
    levels = c(
      "H3b(2)) Ideal Height \n(Level)",
      "H3b(1)) Ideal Height \n(Importance)",
      "H3a(2)) Ideal Age \n(Level)",
      "H3a(1)) Ideal Age \n(Importance)",
      "H2e) Attractiveness",
      "H2d) Kindness-\nSupportiveness",
      "H2c) Education-\nIntelligence",
      "H2b) Confidence-\nAssertiveness",
      "H2a) Financial Security-\nSuccessfulness"
    )
  ))


data_graph2$name = factor(rep(
  c(
    paste0(
      "France (n = ",
      sum(data_included_documented_rescon$country == "France"),
      ")"
    ),
    paste0(
      "Germany (n = ",
      sum(data_included_documented_rescon$country == "Germany"),
      ")"
    ),
    paste0(
      "United States of America (n = ",
      sum(
        data_included_documented_rescon$country == "United States of America"
      ),
      ")"
    ),
    paste0(
      "Mexico (n = ",
      sum(data_included_documented_rescon$country == "Mexico"),
      ")"
    ),
    paste0(
      "Italy (n = ",
      sum(data_included_documented_rescon$country == "Italy"),
      ")"
    ),
    paste0(
      "Brazil (n = ",
      sum(data_included_documented_rescon$country == "Brazil"),
      ")"
    ),
    paste0(
      "Spain (n = ",
      sum(data_included_documented_rescon$country == "Spain"),
      ")"
    )
  ),
  9
))

l_countries2 = levels(data_graph2$name)
data_graph2$name = factor(data_graph2$name,
                          levels = rev(
                            c(
                              l_countries[2],
                              l_countries[3],
                              l_countries[7],
                              l_countries[5],
                              l_countries[4],
                              l_countries[1],
                              l_countries[6]
                            )
                          ))
```

#### Graph Two Multiplot
```{r}
multiplot2 = ggplot(data_graph2, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_pointrange(
    data = data_graph2,
    mapping = aes(
      x = outcome,
      y = mean,
      ymin = conf_low ,
      ymax = conf_high ,
      group = name,
      color = name
    ),
    position = position_dodge(width = 0.4)
  ) +
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 15),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE))
multiplot2
```

##### Save image
```{r}
jpeg(
  "kuschel-figureS2_2.jpeg",
  width = 2*5000,
  height = 2*3000,
  res = 2*300
)
multiplot2
dev.off()
```

#### Combine into one graph
```{r}
data_graph = rbind(
  models_pref_politicalsim_lin_graph,
  models_pref_politicalsim_quad_graph_reg1,
  models_pref_politicalsim_quad_graph_reg2,
  models_pref_ethnicalsim_graph,
  models_pref_religioussim_graph,
  models_pref_level_financially_secure_successful_ambitious_graph,
  models_pref_level_kind_supportive_graph,
  models_pref_level_attractiveness_graph,
  models_pref_level_confident_assertive_graph,
  models_pref_level_intelligence_educated_graph,
  models_imp_age_graph,
  models_ideal_age_rel_graph,
  models_imp_height_graph,
  models_ideal_height_graph
) %>%
  mutate(outcome = factor(
    outcome,
    levels = c(
      "H3b(2)) Ideal Height \n(Level)",
      "H3b(1)) Ideal Height \n(Importance)",
      "H3a(2)) Ideal Age \n(Level)",
      "H3a(1)) Ideal Age \n(Importance)",
      "H2e) Attractiveness",
      "H2d) Kindness-\nSupportiveness",
      "H2c) Education-\nIntelligence",
      "H2b) Confidence-\nAssertiveness",
      "H2a) Financial Security-\nSuccessfulness",
      "H1c) Preferred \nReligious Similarity",
      "H1b) Preferred \nEthnic Similarity",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO >= 3",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO <= 3",
      "H1a(1)) Preferred Political \nSimilarity - Linear Effect"
    )
  ))


data_graph$name = factor(rep(
  c(
    paste0(
      "France (n = ",
      sum(data_included_documented_rescon$country == "France"),
      ")"
    ),
    paste0(
      "Germany (n = ",
      sum(data_included_documented_rescon$country == "Germany"),
      ")"
    ),
    paste0(
      "United States of America (n = ",
      sum(
        data_included_documented_rescon$country == "United States of America"
      ),
      ")"
    ),
    paste0(
      "Mexico (n = ",
      sum(data_included_documented_rescon$country == "Mexico"),
      ")"
    ),
    paste0(
      "Italy (n = ",
      sum(data_included_documented_rescon$country == "Italy"),
      ")"
    ),
    paste0(
      "Brazil (n = ",
      sum(data_included_documented_rescon$country == "Brazil"),
      ")"
    ),
    paste0(
      "Spain (n = ",
      sum(data_included_documented_rescon$country == "Spain"),
      ")"
    )
  ),
  14
))

l_countries = levels(data_graph$name)
data_graph$name = factor(data_graph$name,
                         levels = rev(
                           c(
                             l_countries[2],
                             l_countries[3],
                             l_countries[7],
                             l_countries[5],
                             l_countries[4],
                             l_countries[1],
                             l_countries[6]
                           )
                         ))
```

#### Graph Combined Multiplot {.active}
```{r}
multiplot = ggplot(data_graph, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_pointrange(
    data = data_graph,
    mapping = aes(
      x = outcome,
      y = mean,
      ymin = conf_low ,
      ymax = conf_high ,
      group = name,
      color = name
    ),
    position = position_dodge(width = 0.4)
  ) +
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 15),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE))
multiplot
```

#### Save image
```{r}
jpeg(
  "Exploratory_combined.jpeg",
  width = 5000,
  height = 5000,
  res = 300
)
multiplot
dev.off()
```

### Graph Exploratory Analyses for Regions {.tabset}
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### Inclusion of Data
```{r}
data_included_documented <- data_included_documented %>%
  mutate(region1 = countrycode(country,
                                 origin = "country.name",
                                 destination = "region"),
         region1 = ifelse(country == "Micronesia",
                            "East Asia and Pacific",
                            region1),
         region2 = countrycode(country,
                                 origin = "country.name",
                                 destination = "continent"),
         region2 = ifelse(country == "Micronesia",
                            "Oceania",
                            region2))

table(data_included_documented$region1)
table(data_included_documented$region2)

data_included_documented = data_included_documented %>%
  mutate(region_final = ifelse(region2 == "Africa", "Africa",
                                  ifelse(region2 == "Asia", "Asia",
                                         ifelse(region2 == "Europe", "Europe",
                           ifelse(region2 == "Oceania", "Oceania",
                                  ifelse(region1 == "Latin America & Caribbean",
                                         "Latin and South America",
                                         ifelse(region1 == "North America",
                                                "North America", NA)))))))

data_included_documented <- data_included_documented %>%
  mutate(region_final = ifelse(country %in% c("Afghanistan", "Bahrain", "Iran",
                                                 "Iraq", "Israel", "Jordan", 
                                                 "Kazakhstan", "Kuwait", "Kyrgyzstan",
                                                 "Lebanon", "Pakistan", 
                                                 "Palestinian Territories", "Qatar",
                                                 "Saudi Arabia", "Syria", 
                                                 "Turkey", "Turkmenistan",
                                                 "United Arab Emirates", "Indonesia"),
                                  "Middle East and Central Asia", region_final),
         region_final = ifelse(country %in% c("Armenia", "Burma",
                                                 "China",
                                                 "East Timor (see Timor-Leste)", 
                                                 "Georgia", "Hong Kong", "India",
                                                 "Japan", "Malaysia", "Maldives",
                                                 "Nepal", "Philippines", "Singapore",
                                                 "South Korea", "Sri Lanka", 
                                                 "Taiwan", "Thailand", "Vietnam"), 
                                  "South and East Asia", region_final))

x <- data_included_documented %>% 
  select(country, region_final) %>%
  group_by(country, region_final) %>%
  summarize(n = n()) %>%
  arrange(region_final, country)

write.csv2(x, file = "country_regions.csv")

regions <- x %>%
  group_by(region_final) %>%
  summarize(countries = n(),
            participants = sum(n))

regions
```
We will include all regions with more than 500 participants. This allows us to show effect sizes for a diverse range of regions.

These regions include the following countries (n)

Europe included 41 countries: France (n = 2,013); Germany (n = 1,846); Italy (n = 968); Spain (n = 562); United Kingdom (n = 499); Denmark (n = 395); Switzerland (n = 280); Austria (n = 197); Russia (n = 155); Belgium (n = 102); Ireland (n = 59); Portugal (n = 51); Netherlands (n = 44); Finland (n = 31); Sweden (n = 27); Romania (n = 24); Ukraine (n = 21); Belarus (n = 13); Luxembourg (n = 13); Estonia (n = 12); Norway (n = 12); Czechia (n = 11); Bulgaria (n = 9); Hungary (n = 9); Latvia (n = 8); Poland (n = 7); Andorra (n = 6); Bosnia and Herzegovina (n   =  6); Croatia (n = 6); Iceland (n = 5); Serbia (n = 5); Greece (n = 4); Lithuania (n = 3); Slovakia (n = 3); Slovenia (n = 3); Albania (n = 2); Malta (n = 2); Liechtenstein (n = 1); Macedonia (n = 1); Monaco (n = 1); and Montenegro (n = 1).

Latin and South America included 33 countries: Mexico (n = 1157); Brazil (n = 806); Colombia (n = 387); Argentina (n = 217); Chile (n = 154); Peru (n = 119); Ecuador (n = 102); Venezuela (n = 67); Guatemala (n = 61); Costa Rica (n = 47); Dominican Republic (n = 41); El Salvador (n = 26); Uruguay (n = 24); Bolivia (n = 23); Honduras (n = 18); Nicaragua (n = 15); Panama (n = 15); Paraguay (n = 11); Trinidad and Tobago (n = 8); Jamaica (n = 6); Haiti (n = 4); Guyana (n = 3); Antigua and Barbuda (n = 2); Dominica (n = 2); Saint Lucia (n = 2); Aruba (n = 1); Bahamas (n = 1); Barbados (n = 1); Belize (n = 1); Cuba (n = 1); Grenada (n = 1); Saint Vincent and the Grenadines (n = 1); and Sint Maarten (n = 1).

North America included 2 countries: United States of America (n = 1254) and Canada (n = 338).

South and East Asia included 18 countries: Japan (n = 290); China (n = 90); India (n = 45); Philippines (n = 36); Singapore (n = 26); Malaysia (n = 17); Hong Kong (n = 8); Taiwan (n = 7); South Korea (n = 5); Thailand (n = 4); Georgia (n = 3); Sri Lanka (n = 3); Armenia (n = 2); Vietnam (n = 2); Burma (n = 1); East Timor (n = 1); Maldives (n = 1); and Nepal (n = 1).

Oceania included 5 countries: Australia (n = 133); New Zealand (n = 34); Fiji (n = 1); Marshall Islands (n = 1); and Micronesia (n = 1).

Africa included 26 countries: South Africa (n = 29); Morocco (n = 16); Algeria (n = 7); Tunisia (n = 7); Kenya (n = 6); Nigeria (n = 6); Senegal (n = 5); Namibia (n = 4); Cameroon (n = 3); Egypt (n = 3); Ghana (n = 3); Benin (n = 2); Central African Republic (n = 2); Mali (n = 2); Mauritius (n = 2); Botswana (n = 1); Cote d'Ivoire (n = 1); Ethiopia (n = 1); Guinea-Bissau (n = 1); Madagascar (n = 1); Mauritania (n = 1); South Sudan (n = 1); Swaziland (n = 1); Tanzania (n = 1); Uganda (n = 1); and Zimbabwe (n = 1).

Middle East and Central Asia included 19 countries: Indonesia (n = 18); United Arab Emirates (n = 14); Israel (n = 13); Kazakhstan (n = 9); Pakistan (n = 8); Turkey (n = 8); Iran (n = 7); Saudi Arabia (n = 6); Bahrain (n = 3); Lebanon (n = 3); Jordan (n = 2); Kyrgyzstan (n = 2); Palestinian Territories (n = 2); Qatar (n = 2); Turkmenistan (n = 2); Afghanistan (n = 1); Iraq (n = 1); Kuwait (n = 1); and Syria (n = 1).


We investigated how effects of political orientation on partner preferences differed between four different regions. Namely, these were Europe  (n = 7,417), Latin and South America  (n = 3,325), North America  (n = 1,592), and South and East Asia  (n = 542). We could not include Oceania  (n = 170), Africa  (n = 108), and Middle East and Central Asia  (n = 103) because sample sizes were too small (n < 500) to reach any conclusions.


```{r}
regions = regions %>% filter(participants > 500)
data_included_documented_reg = data_included_documented %>%
  filter(region_final %in% regions$region_final)
```

```{r}
regions_reg =
  data_included_documented_reg %>%
  select(region_final, country) %>%
  table() %>%
  as.data.frame() %>%
  filter(Freq != 0) %>%
  arrange(-Freq)


regions_reg
```

#### Models
##### H1a Preference for similarity in political beliefs
###### H1a(1) Linear Effect
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_politicalsim = data_included_documented_reg_wide %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$pref_politicalsim) ~ scale(.x),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_pref_politicalsim_lin_coef = models_pref_politicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_lin_confint_low = models_pref_politicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_lin_confint_high = models_pref_politicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_lin_graph = left_join(
  left_join(
    models_pref_politicalsim_lin_coef,
    models_pref_politicalsim_lin_confint_low,
    by = "name"
  ),
  models_pref_politicalsim_lin_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1a(1)) Preferred Political \nSimilarity - Linear Effect")
```

###### H1a(2) Quadratic Effect
####### Regression 1 (x <= breaking_point)
```{r}
data_included_documented_reg_wide_reg1 = data_included_documented_reg %>%
  dplyr::filter(political_orientation <= 3) %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')



models_pref_politicalsim_reg1 = data_included_documented_reg_wide_reg1 %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(
      data_included_documented_reg_wide_reg1$pref_politicalsim
    ) ~
      scale(.x),
    data = data_included_documented_reg_wide_reg1
  )) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg1 = models_pref_politicalsim_reg1 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_quad_confint_low_reg1 = models_pref_politicalsim_reg1 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_quad_confint_high_reg1 = models_pref_politicalsim_reg1 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_quad_graph_reg1 = left_join(
  left_join(
    models_pref_politicalsim_quad_coef_reg1,
    models_pref_politicalsim_quad_confint_low_reg1,
    by = "name"
  ),
  models_pref_politicalsim_quad_confint_high_reg1,
  by = "name"
) %>%
  mutate(outcome = "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO <= 3")
```


####### Regression 2 (x >= breaking_point)
```{r}
data_included_documented_reg_wide_reg2 = data_included_documented_reg %>%
  dplyr::filter(political_orientation >= 3) %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_politicalsim_reg2 = data_included_documented_reg_wide_reg2 %>%
  select(-pref_politicalsim) %>%
  map( ~ lm(
    scale(
      data_included_documented_reg_wide_reg2$pref_politicalsim
    ) ~
      scale(.x),
    data = data_included_documented_reg_wide_reg2
  )) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg2 = models_pref_politicalsim_reg2 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_quad_confint_low_reg2 = models_pref_politicalsim_reg2 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_politicalsim_quad_confint_high_reg2 = models_pref_politicalsim_reg2 %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_politicalsim_quad_graph_reg2 = left_join(
  left_join(
    models_pref_politicalsim_quad_coef_reg2,
    models_pref_politicalsim_quad_confint_low_reg2,
    by = "name"
  ),
  models_pref_politicalsim_quad_confint_high_reg2,
  by = "name"
) %>%
  mutate(outcome = "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO >= 3")
```

##### H1b Preference for similarity in ethnicity/race
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_ethnicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_ethnicalsim = data_included_documented_reg_wide %>%
  select(-pref_ethnicalsim) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$pref_ethnicalsim) ~ scale(.x) +
      I(scale(.x) ^ 2),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_pref_ethnicalsim_coef = models_pref_ethnicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_ethnicalsim_confint_low = models_pref_ethnicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_ethnicalsim_confint_high = models_pref_ethnicalsim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_ethnicalsim_graph = left_join(
  left_join(
    models_pref_ethnicalsim_coef,
    models_pref_ethnicalsim_confint_low,
    by = "name"
  ),
  models_pref_ethnicalsim_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1b) Preferred \nEthnic Similarity")
```

##### H1c Preference for similarity in religion
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_religioussim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_religioussim = data_included_documented_reg_wide %>%
  select(-pref_religioussim) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$pref_religioussim) ~ scale(.x) +
      I(scale(.x) ^ 2),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_pref_religioussim_coef = models_pref_religioussim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_religioussim_confint_low = models_pref_religioussim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_religioussim_confint_high = models_pref_religioussim %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_religioussim_graph = left_join(
  left_join(
    models_pref_religioussim_coef,
    models_pref_religioussim_confint_low,
    by = "name"
  ),
  models_pref_religioussim_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H1c) Preferred \nReligious Similarity")
```

##### H2a Preference for the level of financial security- successfulness 
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_level_financially_secure_successful_ambitious, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_financially_secure_successful_ambitious = data_included_documented_reg_wide %>%
  select(-pref_level_financially_secure_successful_ambitious) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$pref_level_financially_secure_successful_ambitious ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_financially_secure_successful_ambitious_coef = models_pref_level_financially_secure_successful_ambitious %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_financially_secure_successful_ambitious_confint_low = models_pref_level_financially_secure_successful_ambitious %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_financially_secure_successful_ambitious_confint_high = models_pref_level_financially_secure_successful_ambitious %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_financially_secure_successful_ambitious_graph = left_join(
  left_join(
    models_pref_level_financially_secure_successful_ambitious_coef,
    models_pref_level_financially_secure_successful_ambitious_confint_low,
    by = "name"
  ),
  models_pref_level_financially_secure_successful_ambitious_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2a) Financial Security-\nSuccessfulness")
```

##### H2b Preference for the level of confidence-assertiveness
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_level_confident_assertive, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_confident_assertive = data_included_documented_reg_wide %>%
  select(-pref_level_confident_assertive) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$pref_level_confident_assertive ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_confident_assertive_coef = models_pref_level_confident_assertive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_confident_assertive_confint_low = models_pref_level_confident_assertive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_confident_assertive_confint_high = models_pref_level_confident_assertive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_confident_assertive_graph = left_join(
  left_join(
    models_pref_level_confident_assertive_coef,
    models_pref_level_confident_assertive_confint_low,
    by = "name"
  ),
  models_pref_level_confident_assertive_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2b) Confidence-\nAssertiveness")
```

##### H2c Preference for the level of education-intelligence
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_level_intelligence_educated, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_intelligence_educated = data_included_documented_reg_wide %>%
  select(-pref_level_intelligence_educated) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$pref_level_intelligence_educated ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_intelligence_educated_coef = models_pref_level_intelligence_educated %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_intelligence_educated_confint_low = models_pref_level_intelligence_educated %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_intelligence_educated_confint_high = models_pref_level_intelligence_educated %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_intelligence_educated_graph = left_join(
  left_join(
    models_pref_level_intelligence_educated_coef,
    models_pref_level_intelligence_educated_confint_low,
    by = "name"
  ),
  models_pref_level_intelligence_educated_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2c) Education-\nIntelligence")
```

##### H2d Preference for the level of kindness-supportiveness
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_level_kind_supportive, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_kind_supportive = data_included_documented_reg_wide %>%
  select(-pref_level_kind_supportive) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$pref_level_kind_supportive ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_kind_supportive_coef = models_pref_level_kind_supportive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_kind_supportive_confint_low = models_pref_level_kind_supportive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_kind_supportive_confint_high = models_pref_level_kind_supportive %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_kind_supportive_graph = left_join(
  left_join(
    models_pref_level_kind_supportive_coef,
    models_pref_level_kind_supportive_confint_low,
    by = "name"
  ),
  models_pref_level_kind_supportive_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2d) Kindness-\nSupportiveness")
```

##### H2e Preference for the level of attractiveness
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    pref_level_attractiveness, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_attractiveness = data_included_documented_reg_wide %>%
  select(-pref_level_attractiveness) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$pref_level_attractiveness ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_pref_level_attractiveness_coef = models_pref_level_attractiveness %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_attractiveness_confint_low = models_pref_level_attractiveness %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_pref_level_attractiveness_confint_high = models_pref_level_attractiveness %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_pref_level_attractiveness_graph = left_join(
  left_join(
    models_pref_level_attractiveness_coef,
    models_pref_level_attractiveness_confint_low,
    by = "name"
  ),
  models_pref_level_attractiveness_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H2e) Attractiveness")
```


##### H3a(1) Ideal age (importance)
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(imp_age, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_imp_age = data_included_documented_reg_wide %>%
  select(-imp_age) %>%
  map(
    ~ lm(data_included_documented_reg_wide$imp_age ~ .x,
         data = data_included_documented_reg_wide)
  ) %>%
  map(lm.beta)

models_imp_age_coef = models_imp_age %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_age_confint_low = models_imp_age %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_imp_age_confint_high = models_imp_age %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_imp_age_graph = left_join(
  left_join(models_imp_age_coef,
            models_imp_age_confint_low,
            by = "name"),
  models_imp_age_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3a(1)) Ideal Age \n(Importance)")
```

##### H3a(2) Ideal age (level)
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(ideal_age_rel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_ideal_age_rel = data_included_documented_reg_wide %>%
  select(-ideal_age_rel) %>%
  map(
    ~ lm(
      data_included_documented_reg_wide$ideal_age_rel ~ .x,
      data = data_included_documented_reg_wide
    )
  ) %>%
  map(lm.beta)

models_ideal_age_rel_coef = models_ideal_age_rel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_age_rel_confint_low = models_ideal_age_rel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_ideal_age_rel_confint_high = models_ideal_age_rel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_ideal_age_rel_graph = left_join(
  left_join(
    models_ideal_age_rel_coef,
    models_ideal_age_rel_confint_low,
    by = "name"
  ),
  models_ideal_age_rel_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3a(2)) Ideal Age \n(Level)")
```

##### H3b(1) Ideal height (importance)
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(imp_height, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_imp_height = data_included_documented_reg_wide %>%
  select(-imp_height) %>%
  map(
    ~ lm(data_included_documented_reg_wide$imp_height ~ .x,
         data = data_included_documented_reg_wide)
  ) %>%
  map(lm.beta)

models_imp_height_coef = models_imp_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_height_confint_low = models_imp_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_imp_height_confint_high = models_imp_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_imp_height_graph = left_join(
  left_join(models_imp_height_coef,
            models_imp_height_confint_low,
            by = "name"),
  models_imp_height_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3b(1)) Ideal Height \n(Importance)")
```

##### H3b(2) Ideal height (level)
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(ideal_height, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_ideal_height = data_included_documented_reg_wide %>%
  select(-ideal_height) %>%
  map(
    ~ lm(data_included_documented_reg_wide$ideal_height ~ .x,
         data = data_included_documented_reg_wide)
  ) %>%
  map(lm.beta)

models_ideal_height_coef = models_ideal_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_height_confint_low = models_ideal_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_ideal_height_confint_high = models_ideal_height %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_ideal_height_graph = left_join(
  left_join(
    models_ideal_height_coef,
    models_ideal_height_confint_low,
    by = "name"
  ),
  models_ideal_height_confint_high,
  by = "name"
) %>%
  mutate(outcome = "H3b(2)) Ideal Height \n(Level)")
```

#### Combine Graph One
```{r}
data_graph = rbind(
  models_pref_politicalsim_lin_graph,
  models_pref_politicalsim_quad_graph_reg1,
  models_pref_politicalsim_quad_graph_reg2,
  models_pref_ethnicalsim_graph,
  models_pref_religioussim_graph
) %>%
  mutate(outcome = factor(
    outcome,
    levels = c(
      "H1c) Preferred \nReligious Similarity",
      "H1b) Preferred \nEthnic Similarity",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO >= 3",
      "H1a(2)) Preferred Political \nSimilarity - Quadratic Effect \nPO <= 3",
      "H1a(1)) Preferred Political \nSimilarity - Linear Effect"
    )
  ))


data_graph$name = factor(rep(
  c(
    paste0(
      "Europe (n = ",
      sum(data_included_documented_reg$region_final == "Europe"),
      ")"
    ),
    paste0(
      "South and East Asia (n = ",
      sum(data_included_documented_reg$region_final == "South and East Asia"),
      ")"
    ),
    paste0(
      "Latin and South America (n = ",
      sum(
        data_included_documented_reg$region_final == "Latin and South America"
      ),
      ")"
    ),
    paste0(
      "North America (n = ",
      sum(data_included_documented_reg$region_final == "North America"),
      ")"
    )
  ),
  5
))

l_countries = levels(data_graph$name)
data_graph$name = factor(data_graph$name,
                         levels = rev(
                           c(
                             l_countries[1],
                             l_countries[2],
                             l_countries[3],
                             l_countries[4])
                         ))
```

#### Graph One Multiplot {.active}
```{r}
multiplot = ggplot(data_graph, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_pointrange(
    data = data_graph,
    mapping = aes(
      x = outcome,
      y = mean,
      ymin = conf_low ,
      ymax = conf_high ,
      group = name,
      color = name
    ),
    position = position_dodge(width = 0.4)
  ) +
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 15),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE))
multiplot
```

##### Save image
```{r}
jpeg(
  "kuschel-figure_exploratiory_regions1.jpeg",
  width = 2*5000,
  height = 2*3000,
  res = 600
)
multiplot
dev.off()
```

#### Combine Graph Two
```{r}
data_graph2 = rbind(
  models_pref_level_financially_secure_successful_ambitious_graph,
  models_pref_level_kind_supportive_graph,
  models_pref_level_attractiveness_graph,
  models_pref_level_confident_assertive_graph,
  models_pref_level_intelligence_educated_graph,
  models_imp_age_graph,
  models_ideal_age_rel_graph,
  models_imp_height_graph,
  models_ideal_height_graph
) %>%
  mutate(outcome = factor(
    outcome,
    levels = c(
      "H3b(2)) Ideal Height \n(Level)",
      "H3b(1)) Ideal Height \n(Importance)",
      "H3a(2)) Ideal Age \n(Level)",
      "H3a(1)) Ideal Age \n(Importance)",
      "H2e) Attractiveness",
      "H2d) Kindness-\nSupportiveness",
      "H2c) Education-\nIntelligence",
      "H2b) Confidence-\nAssertiveness",
      "H2a) Financial Security-\nSuccessfulness"
    )
  ))



data_graph2$name = factor(rep(
  c(
    paste0(
      "Europe (n = ",
      sum(data_included_documented_reg$region_final == "Europe"),
      ")"
    ),
    paste0(
      "South and East Asia (n = ",
      sum(data_included_documented_reg$region_final == "South and East Asia"),
      ")"
    ),
    paste0(
      "Latin and South America (n = ",
      sum(
        data_included_documented_reg$region_final == "Latin and South America"
      ),
      ")"
    ),
    paste0(
      "North America (n = ",
      sum(data_included_documented_reg$region_final == "North America"),
      ")"
    )
  ),
  9
))

l_countries = levels(data_graph2$name)
data_graph2$name = factor(data_graph2$name,
                         levels = rev(
                           c(
                             l_countries[1],
                             l_countries[2],
                             l_countries[3],
                             l_countries[4])
                         ))
```

#### Graph Two Multiplot
```{r}
multiplot2 = ggplot(data_graph2, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_pointrange(
    data = data_graph2,
    mapping = aes(
      x = outcome,
      y = mean,
      ymin = conf_low ,
      ymax = conf_high ,
      group = name,
      color = name
    ),
    position = position_dodge(width = 0.4)
  ) +
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 20),
    axis.text.x = element_text(size = 15),
    axis.text.y = element_text(size = 15),
    legend.text = element_text(size = 15),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  guides(col = guide_legend(reverse = TRUE))
multiplot2
```

##### Save image
```{r}
jpeg(
  "kuschel-figure_exploratiory_regions2.jpeg",
  width = 2*5000,
  height = 2*3000,
  res = 2*300
)
multiplot2
dev.off()
```



### Figure S3 {.tabset}
Effect size estimates of linear effects of political orientation on preferred relationship options for main analyses and one robustness analysis.

#### Main Analyses
```{r}
model_interest_single <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_single ~ political_orientation + (1 + political_orientation |
                                                   country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_sexrel <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_sexrel ~ political_orientation + (1 + political_orientation |
                                                   country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_nonmonrel <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_nonmonrel ~ political_orientation + (1 + political_orientation |
                                                      country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_monrel <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_monrel ~ political_orientation + (1 + political_orientation |
                                                   country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```

#### Robustness Analyses
Additionally including age as a predictor
```{r}
model_interest_single_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_single ~ political_orientation + age + (1 + political_orientation |
                                                         country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_sexrel_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_sexrel ~ political_orientation + age + (1 + political_orientation |
                                                         country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_nonmonrel_robustcheck <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_nonmonrel ~ political_orientation + age + (1 + political_orientation |
                                                            country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))

model_interest_monrel_robustcheck  <-
  as.data.frame(standardize_parameters(
    lmer(
      interest_monrel ~ political_orientation + age + (1 + political_orientation |
                                                         country),
      data = data_included_documented,
      control = lmerControl(optimizer = "bobyqa")
    ),
    method = "basic",
    ci = 0.997
  ))
```


#### Combine {.active}
```{r}
all_models_interest <- rbind(
  model_interest_single,
  model_interest_sexrel,
  model_interest_nonmonrel,
  model_interest_monrel,
  model_interest_single_robustcheck,
  model_interest_sexrel_robustcheck,
  model_interest_nonmonrel_robustcheck,
  model_interest_monrel_robustcheck
) %>%
  filter(Parameter == "political_orientation")

all_models_interest$hypothesis = factor(
  rep(
    c(
      "S1) Interest in Single Life",
      "S2) Interest in Sexual,\nNon-Romantic Relationships",
      "S3) Interest in \nNon-Monogamous Relationships",
      "S4) Interest in \nMonogamous Relationships"
    ),
    2
  ),
  levels = c(
    "S4) Interest in \nMonogamous Relationships",
    "S3) Interest in \nNon-Monogamous Relationships",
    "S2) Interest in Sexual,\nNon-Romantic Relationships",
    "S1) Interest in Single Life"
  )
)

all_models_interest$type = factor(c(rep("Main Analyses", 4), rep("Robustness Analyses", 4)),
                                  levels = c("Robustness Analyses", "Main Analyses"))


multiplot = ggplot(all_models_interest, aes(hypothesis, Std_Coefficient)) +
  geom_pointrange(
    data = all_models_interest,
    mapping = aes(
      x = hypothesis,
      y = Std_Coefficient,
      ymin = CI_low,
      ymax = CI_high,
      group = type,
      color = type
    ),
    position = position_dodge(width = 0.3)
  ) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05,
             linetype = "dotted",
             linewidth = 1.5) +
  geom_hline(yintercept = .05,
             linetype = "dotted",
             linewidth = 1.5) +
  coord_flip() +
  apatheme +
  scale_colour_manual(values = c("#86BDDB", "#084594"))  +
  labs(y = "Linear Effect of Political Orientation", x = "Outcome") +
  theme(
    text = element_text(size = 15),
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 10),
    legend.text = element_text(size = 10),
    plot.title = element_text(hjust = 0),
    legend.background = element_rect(
      fill = "grey",
      linewidth = 0.5,
      linetype = "solid"
    )
  ) +
  #theme(legend.position=c(0.8, 0.8)) +
  guides(col = guide_legend(reverse = TRUE))
multiplot
```

#### Save image
```{r}
jpeg("kuschel-figureS3.jpeg",
     width = 6300,
     height = 4200,
     res = 600)
multiplot
dev.off()
```

### Figure S4
#### Inclusion of data
```{r}
countries = as.data.frame(table(data_included_documented$country)) %>%
  arrange(-Freq)

countries
```
We will include all countries with more than 500 participants. This allows us to show effect sizes for a diverse range of countries.
Diversity of countries is indicated by:

* location: European (France, Germany, Italy, Spain); North American (United States of America); South American (Mexico, Brazil)
* language: French (France); German (Germany); English (United States of America); Spanish (Mexico, Spain); Italian (Italy); Portuguese (Brazil)
* culture: Western (France, Germany, Italy, Spain, United States of America); Non-Western (Mexico, Brazil)

Sample sizes of other countries are too small (n < 500) to reach any conclusions.

```{r}
seven_countries = countries %>% filter(Freq > 500)
data_included_documented_rescon = data_included_documented %>%
  filter(country %in% seven_countries$Var1)
```

```{r}
countries_rescon = as.data.frame(table(data_included_documented_rescon$country)) %>%
  arrange(-Freq)
countries_rescon
```

#### Models
##### S1 Interest in Single Life
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_single, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_single = data_included_documented_rescon_wide %>%
  select(-interest_single) %>%
  map(~lm(data_included_documented_rescon_wide$interest_single ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_single_coef = models_interest_single %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_single_confint_low = models_interest_single %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name)-7))

models_interest_single_confint_high = models_interest_single %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name)-8))

models_interest_single_graph = left_join(left_join(models_interest_single_coef,
                                         models_interest_single_confint_low,
                                         by = "name"),
                                         models_interest_single_confint_high,
                                         by = "name") %>%
  mutate(outcome = "S1) Interest in Single Life")
```

#####  S2 Interest in sexual, non-romantic relationships
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_sexrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_sexrel = data_included_documented_rescon_wide %>%
  select(-interest_sexrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_sexrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_sexrel_coef = models_interest_sexrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_sexrel_confint_low = models_interest_sexrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name)-7))

models_interest_sexrel_confint_high = models_interest_sexrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name)-8))

models_interest_sexrel_graph = left_join(left_join(models_interest_sexrel_coef,
                                         models_interest_sexrel_confint_low,
                                         by = "name"),
                                         models_interest_sexrel_confint_high,
                                         by = "name") %>%
  mutate(outcome = "S2) Interest in Sexual,\nNon-Romantic Relationships")
```

##### S3 Interest in non-monogamous relationships
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_nonmonrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_nonmonrel = data_included_documented_rescon_wide %>%
  select(-interest_nonmonrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_nonmonrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_nonmonrel_coef = models_interest_nonmonrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_nonmonrel_confint_low = models_interest_nonmonrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name)-7))

models_interest_nonmonrel_confint_high = models_interest_nonmonrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name)-8))

models_interest_nonmonrel_graph = left_join(left_join(models_interest_nonmonrel_coef,
                                         models_interest_nonmonrel_confint_low,
                                         by = "name"),
                                         models_interest_nonmonrel_confint_high,
                                         by = "name") %>%
  mutate(outcome = "S3) Interest in \nNon-Monogamous Relationships")
```

##### S4 Interest in monogamous relationships
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_monrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_monrel = data_included_documented_rescon_wide %>%
  select(-interest_monrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_monrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_monrel_coef = models_interest_monrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_monrel_confint_low = models_interest_monrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name)-7))

models_interest_monrel_confint_high = models_interest_monrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name)-8))

models_interest_monrel_graph = left_join(left_join(models_interest_monrel_coef,
                                         models_interest_monrel_confint_low,
                                         by = "name"),
                                         models_interest_monrel_confint_high,
                                         by = "name") %>%
  mutate(outcome = "S4) Interest in \nMonogamous Relationships")
```

#### Combine Graph One
```{r}
data_graph = rbind(models_interest_single_graph, models_interest_sexrel_graph,
                   models_interest_nonmonrel_graph, models_interest_monrel_graph
                   ) %>%
  mutate(outcome = factor(outcome,
                          levels = c(
                                     "S4) Interest in \nMonogamous Relationships",
                                     "S3) Interest in \nNon-Monogamous Relationships",
                                     "S2) Interest in Sexual,\nNon-Romantic Relationships",
                                   "S1) Interest in Single Life")))


data_graph$name = factor(rep(c(
  paste0("France (n = ",
         sum(data_included_documented_rescon$country == "France"), ")"),
  paste0("Germany (n = ",
         sum(data_included_documented_rescon$country == "Germany"), ")"),
  paste0("United States of America (n = ",
         sum(data_included_documented_rescon$country == "United States of America"),
         ")"),
  paste0("Mexico (n = ",
         sum(data_included_documented_rescon$country == "Mexico"), ")"),
  paste0("Italy (n = ",
         sum(data_included_documented_rescon$country == "Italy"), ")"),
  paste0("Brazil (n = ",
         sum(data_included_documented_rescon$country == "Brazil"), ")"),
  paste0("Spain (n = ",
         sum(data_included_documented_rescon$country == "Spain"), ")")),
  4))

l_countries = levels(data_graph$name)
data_graph$name = factor(data_graph$name,
                         levels = rev(c(l_countries[2], l_countries[3],
                                    l_countries[7], l_countries[5],
                                    l_countries[4], l_countries[1],
                                    l_countries[6])))
```

#### Graph One Multiplot {.active}
```{r}
multiplot = ggplot(data_graph, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05, linetype = "dotted", linewidth = 1.5) + 
  geom_hline(yintercept = .05, linetype = "dotted", linewidth = 1.5) +
  geom_pointrange(data = data_graph,
                  mapping = aes(x = outcome, y = mean,
                                ymin = conf_low , ymax = conf_high , group = name,
                                color = name),
                  position = position_dodge(width=0.4)) + 
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Right-Wing Political Orientation", x = "Outcome") +
  theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
        axis.text.y = element_text(size = 15),
        legend.text = element_text(size = 15),
        plot.title = element_text(hjust = 0),
        legend.background = element_rect(fill="grey", 
                                         size=0.5, linetype="solid")) +
  guides(col = guide_legend(reverse= TRUE))
multiplot
```

##### Save image
```{r}
jpeg("kuschel-figureS4.jpeg", width = 2*5000, height = 2*3000, res=600)
multiplot
dev.off()
```

### Graph Exploratory Analyses for Regions {.tabset}
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### Inclusion of Data
```{r}
data_included_documented <- data_included_documented %>%
  mutate(region1 = countrycode(country,
                                 origin = "country.name",
                                 destination = "region"),
         region1 = ifelse(country == "Micronesia",
                            "East Asia and Pacific",
                            region1),
         region2 = countrycode(country,
                                 origin = "country.name",
                                 destination = "continent"),
         region2 = ifelse(country == "Micronesia",
                            "Oceania",
                            region2))

table(data_included_documented$region1)
table(data_included_documented$region2)

data_included_documented = data_included_documented %>%
  mutate(region_final = ifelse(region2 == "Africa", "Africa",
                                  ifelse(region2 == "Asia", "Asia",
                                         ifelse(region2 == "Europe", "Europe",
                           ifelse(region2 == "Oceania", "Oceania",
                                  ifelse(region1 == "Latin America & Caribbean",
                                         "Latin and South America",
                                         ifelse(region1 == "North America",
                                                "North America", NA)))))))

data_included_documented <- data_included_documented %>%
  mutate(region_final = ifelse(country %in% c("Afghanistan", "Bahrain", "Iran",
                                                 "Iraq", "Israel", "Jordan", 
                                                 "Kazakhstan", "Kuwait", "Kyrgyzstan",
                                                 "Lebanon", "Pakistan", 
                                                 "Palestinian Territories", "Qatar",
                                                 "Saudi Arabia", "Syria", 
                                                 "Turkey", "Turkmenistan",
                                                 "United Arab Emirates", "Indonesia"),
                                  "Middle East and Central Asia", region_final),
         region_final = ifelse(country %in% c("Armenia", "Burma",
                                                 "China",
                                                 "East Timor (see Timor-Leste)", 
                                                 "Georgia", "Hong Kong", "India",
                                                 "Japan", "Malaysia", "Maldives",
                                                 "Nepal", "Philippines", "Singapore",
                                                 "South Korea", "Sri Lanka", 
                                                 "Taiwan", "Thailand", "Vietnam"), 
                                  "South and East Asia", region_final))

x <- data_included_documented %>% 
  select(country, region_final) %>%
  group_by(country, region_final) %>%
  summarize(n = n()) %>%
  arrange(region_final, country)

write.csv2(x, file = "country_regions.csv")

regions <- x %>%
  group_by(region_final) %>%
  summarize(countries = n(),
            participants = sum(n))

regions
```
We will include all regions with more than 500 participants. This allows us to show effect sizes for a diverse range of regions.

These regions include the following countries (n)

Europe included 41 countries: France (n = 2,013); Germany (n = 1,846); Italy (n = 968); Spain (n = 562); United Kingdom (n = 499); Denmark (n = 395); Switzerland (n = 280); Austria (n = 197); Russia (n = 155); Belgium (n = 102); Ireland (n = 59); Portugal (n = 51); Netherlands (n = 44); Finland (n = 31); Sweden (n = 27); Romania (n = 24); Ukraine (n = 21); Belarus (n = 13); Luxembourg (n = 13); Estonia (n = 12); Norway (n = 12); Czechia (n = 11); Bulgaria (n = 9); Hungary (n = 9); Latvia (n = 8); Poland (n = 7); Andorra (n = 6); Bosnia and Herzegovina (n   =  6); Croatia (n = 6); Iceland (n = 5); Serbia (n = 5); Greece (n = 4); Lithuania (n = 3); Slovakia (n = 3); Slovenia (n = 3); Albania (n = 2); Malta (n = 2); Liechtenstein (n = 1); Macedonia (n = 1); Monaco (n = 1); and Montenegro (n = 1).

Latin and South America included 33 countries: Mexico (n = 1157); Brazil (n = 806); Colombia (n = 387); Argentina (n = 217); Chile (n = 154); Peru (n = 119); Ecuador (n = 102); Venezuela (n = 67); Guatemala (n = 61); Costa Rica (n = 47); Dominican Republic (n = 41); El Salvador (n = 26); Uruguay (n = 24); Bolivia (n = 23); Honduras (n = 18); Nicaragua (n = 15); Panama (n = 15); Paraguay (n = 11); Trinidad and Tobago (n = 8); Jamaica (n = 6); Haiti (n = 4); Guyana (n = 3); Antigua and Barbuda (n = 2); Dominica (n = 2); Saint Lucia (n = 2); Aruba (n = 1); Bahamas (n = 1); Barbados (n = 1); Belize (n = 1); Cuba (n = 1); Grenada (n = 1); Saint Vincent and the Grenadines (n = 1); and Sint Maarten (n = 1).

North America included 2 countries: United States of America (n = 1254) and Canada (n = 338).

South and East Asia included 18 countries: Japan (n = 290); China (n = 90); India (n = 45); Philippines (n = 36); Singapore (n = 26); Malaysia (n = 17); Hong Kong (n = 8); Taiwan (n = 7); South Korea (n = 5); Thailand (n = 4); Georgia (n = 3); Sri Lanka (n = 3); Armenia (n = 2); Vietnam (n = 2); Burma (n = 1); East Timor (n = 1); Maldives (n = 1); and Nepal (n = 1).

Oceania included 5 countries: Australia (n = 133); New Zealand (n = 34); Fiji (n = 1); Marshall Islands (n = 1); and Micronesia (n = 1).

Africa included 26 countries: South Africa (n = 29); Morocco (n = 16); Algeria (n = 7); Tunisia (n = 7); Kenya (n = 6); Nigeria (n = 6); Senegal (n = 5); Namibia (n = 4); Cameroon (n = 3); Egypt (n = 3); Ghana (n = 3); Benin (n = 2); Central African Republic (n = 2); Mali (n = 2); Mauritius (n = 2); Botswana (n = 1); Cote d'Ivoire (n = 1); Ethiopia (n = 1); Guinea-Bissau (n = 1); Madagascar (n = 1); Mauritania (n = 1); South Sudan (n = 1); Swaziland (n = 1); Tanzania (n = 1); Uganda (n = 1); and Zimbabwe (n = 1).

Middle East and Central Asia included 19 countries: Indonesia (n = 18); United Arab Emirates (n = 14); Israel (n = 13); Kazakhstan (n = 9); Pakistan (n = 8); Turkey (n = 8); Iran (n = 7); Saudi Arabia (n = 6); Bahrain (n = 3); Lebanon (n = 3); Jordan (n = 2); Kyrgyzstan (n = 2); Palestinian Territories (n = 2); Qatar (n = 2); Turkmenistan (n = 2); Afghanistan (n = 1); Iraq (n = 1); Kuwait (n = 1); and Syria (n = 1).


We investigated how effects of political orientation on partner preferences differed between four different regions. Namely, these were Europe  (n = 7,417), Latin and South America  (n = 3,325), North America  (n = 1,592), and South and East Asia  (n = 542). We could not include Oceania  (n = 170), Africa  (n = 108), and Middle East and Central Asia  (n = 103) because sample sizes were too small (n < 500) to reach any conclusions.


```{r}
regions = regions %>% filter(participants > 500)
data_included_documented_reg = data_included_documented %>%
  filter(region_final %in% regions$region_final)
```

```{r}
regions_reg =
  data_included_documented_reg %>%
  select(region_final, country) %>%
  table() %>%
  as.data.frame() %>%
  filter(Freq != 0) %>%
  arrange(-Freq)


regions_reg
```

#### Models
##### S1 Interest in Single Life
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    interest_single, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_single = data_included_documented_reg_wide %>%
  select(-interest_single) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$interest_single) ~ scale(.x),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_interest_single_coef = models_interest_single %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_single_confint_low = models_interest_single %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_interest_single_confint_high = models_interest_single %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_interest_single_graph = left_join(
  left_join(
    models_interest_single_coef,
    models_interest_single_confint_low,
    by = "name"
  ),
  models_interest_single_confint_high,
  by = "name"
) %>%
  mutate(outcome = "S1) Interest in Single Life")
```

#####  S2 Interest in sexual, non-romantic relationships
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    interest_sexrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_sexrel = data_included_documented_reg_wide %>%
  select(-interest_sexrel) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$interest_sexrel) ~ scale(.x),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_interest_sexrel_coef = models_interest_sexrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_sexrel_confint_low = models_interest_sexrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_interest_sexrel_confint_high = models_interest_sexrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_interest_sexrel_graph = left_join(
  left_join(
    models_interest_sexrel_coef,
    models_interest_sexrel_confint_low,
    by = "name"
  ),
  models_interest_sexrel_confint_high,
  by = "name"
) %>%
  mutate(outcome = "S2) Interest in Sexual,\nNon-Romantic Relationships")
```

##### S3 Interest in non-monogamous relationships
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    interest_nonmonrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_nonmonrel = data_included_documented_reg_wide %>%
  select(-interest_nonmonrel) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$interest_nonmonrel) ~ scale(.x),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_interest_nonmonrel_coef = models_interest_nonmonrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_nonmonrel_confint_low = models_interest_nonmonrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_interest_nonmonrel_confint_high = models_interest_nonmonrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_interest_nonmonrel_graph = left_join(
  left_join(
    models_interest_nonmonrel_coef,
    models_interest_nonmonrel_confint_low,
    by = "name"
  ),
  models_interest_nonmonrel_confint_high,
  by = "name"
) %>%
  mutate(outcome = "S3) Interest in \nNon-Monogamous Relationships")
```

##### S4 Interest in monogamous relationships
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(
    interest_monrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_monrel = data_included_documented_reg_wide %>%
  select(-interest_monrel) %>%
  map( ~ lm(
    scale(data_included_documented_reg_wide$interest_monrel) ~ scale(.x),
    data = data_included_documented_reg_wide
  )) %>%
  map(lm.beta)

models_interest_monrel_coef = models_interest_monrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_monrel_confint_low = models_interest_monrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "0.15")) %>%
  rename(conf_low = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 7))

models_interest_monrel_confint_high = models_interest_monrel %>%
  map(confint, level = 0.997)  %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname  == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  filter(str_detect(name, "99.85")) %>%
  rename(conf_high = value) %>%
  mutate(name = substr(name, 1, nchar(name) - 8))

models_interest_monrel_graph = left_join(
  left_join(
    models_interest_monrel_coef,
    models_interest_monrel_confint_low,
    by = "name"
  ),
  models_interest_monrel_confint_high,
  by = "name"
) %>%
  mutate(outcome = "S4) Interest in \nMonogamous Relationships")
```

#### Combine Graph One
```{r}
data_graph = rbind(models_interest_single_graph, models_interest_sexrel_graph,
                   models_interest_nonmonrel_graph, models_interest_monrel_graph
) %>%
  mutate(outcome = factor(outcome,
                          levels = c(
                            "S4) Interest in \nMonogamous Relationships",
                            "S3) Interest in \nNon-Monogamous Relationships",
                            "S2) Interest in Sexual,\nNon-Romantic Relationships",
                            "S1) Interest in Single Life")))


data_graph$name = factor(rep(
  c(
    paste0(
      "Europe (n = ",
      sum(data_included_documented_reg$region_final == "Europe"),
      ")"
    ),
    paste0(
      "South and East Asia (n = ",
      sum(data_included_documented_reg$region_final == "South and East Asia"),
      ")"
    ),
    paste0(
      "Latin and South America (n = ",
      sum(
        data_included_documented_reg$region_final == "Latin and South America"
      ),
      ")"
    ),
    paste0(
      "North America (n = ",
      sum(data_included_documented_reg$region_final == "North America"),
      ")"
    )
  ),
  4
))

l_countries = levels(data_graph$name)
data_graph$name = factor(data_graph$name,
                         levels = rev(
                           c(
                             l_countries[1],
                             l_countries[2],
                             l_countries[3],
                             l_countries[4])
                         ))
```

#### Graph One Multiplot {.active}
```{r}
multiplot = ggplot(data_graph, aes(outcome, mean)) +
  geom_hline(yintercept = 0, linewidth = 1.5)  +
  geom_hline(yintercept = -.05, linetype = "dotted", linewidth = 1.5) + 
  geom_hline(yintercept = .05, linetype = "dotted", linewidth = 1.5) +
  geom_pointrange(data = data_graph,
                  mapping = aes(x = outcome, y = mean,
                                ymin = conf_low , ymax = conf_high , group = name,
                                color = name),
                  position = position_dodge(width=0.4)) + 
  coord_flip() +
  apatheme +
  scale_colour_brewer(palette = "Dark2")  +
  labs(y = "Effect of Right-Wing Political Orientation", x = "Outcome") +
  theme(text = element_text(size=20), axis.text.x = element_text(size = 15),
        axis.text.y = element_text(size = 15),
        legend.text = element_text(size = 15),
        plot.title = element_text(hjust = 0),
        legend.background = element_rect(fill="grey", 
                                         size=0.5, linetype="solid")) +
  guides(col = guide_legend(reverse= TRUE))
multiplot
```

##### Save image
```{r}
jpeg("kuschel-figure_exploratiory_regions3.jpeg", width = 2*5000, height = 2*3000, res=600)
multiplot
dev.off()
```
