==> 10_analyses_exploratory_regions.Rmd <==
---
title: <font color="#66C2A5">Exploratory Analyses Regions</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## {.tabset}

### Library
```{r Library}
library(formr)
library(effects)
library(effectsize)
library(lme4)
library(sjstats)
library(lmerTest)
library(ggplot2)
library(tidyr)
library(ggpubr)
library(RColorBrewer)
library(coefplot)
library(tibble)
library(purrr) # for running multiple regression
library(broom)
library(mvmeta)
library(lm.beta)
library(dplyr)
library(stringr)
library(tidyr)
library(knitr)
library(countrycode)

apatheme = theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.line = element_line(),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5))
```

### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

### Inclusion of Data
```{r}
data_included_documented <- data_included_documented %>%
  mutate(region1 = countrycode(country,
                                 origin = "country.name",
                                 destination = "region"),
         region1 = ifelse(country == "Micronesia",
                            "East Asia and Pacific",
                            region1),
         region2 = countrycode(country,
                                 origin = "country.name",
                                 destination = "continent"),
         region2 = ifelse(country == "Micronesia",
                            "Oceania",
                            region2))

table(data_included_documented$region1)
table(data_included_documented$region2)

data_included_documented = data_included_documented %>%
  mutate(region_final = ifelse(region2 == "Africa", "Africa",
                                  ifelse(region2 == "Asia", "Asia",
                                         ifelse(region2 == "Europe", "Europe",
                           ifelse(region2 == "Oceania", "Oceania",
                                  ifelse(region1 == "Latin America & Caribbean",
                                         "Latin and South America",
                                         ifelse(region1 == "North America",
                                                "North America", NA)))))))

data_included_documented <- data_included_documented %>%
  mutate(region_final = ifelse(country %in% c("Afghanistan", "Bahrain", "Iran",
                                                 "Iraq", "Israel", "Jordan", 
                                                 "Kazakhstan", "Kuwait", "Kyrgyzstan",
                                                 "Lebanon", "Pakistan", 
                                                 "Palestinian Territories", "Qatar",
                                                 "Saudi Arabia", "Syria", 
                                                 "Turkey", "Turkmenistan",
                                                 "United Arab Emirates", "Indonesia"),
                                  "Middle East and Central Asia", region_final),
         region_final = ifelse(country %in% c("Armenia", "Burma",
                                                 "China",
                                                 "East Timor (see Timor-Leste)", 
                                                 "Georgia", "Hong Kong", "India",
                                                 "Japan", "Malaysia", "Maldives",
                                                 "Nepal", "Philippines", "Singapore",
                                                 "South Korea", "Sri Lanka", 
                                                 "Taiwan", "Thailand", "Vietnam"), 
                                  "South and East Asia", region_final))

x <- data_included_documented %>% 
  select(country, region_final) %>%
  group_by(country, region_final) %>%
  summarize(n = n()) %>%
  arrange(region_final, country)

write.csv2(x, file = "country_regions.csv")

regions <- x %>%
  group_by(region_final) %>%
  summarize(countries = n(),
            participants = sum(n))

regions
```
We will include all regions with more than 500 participants. This allows us to show effect sizes for a diverse range of regions.

These regions include the following countries (n)

Europe included 41 countries: France (n = 2,013); Germany (n = 1,846); Italy (n = 968); Spain (n = 562); United Kingdom (n = 499); Denmark (n = 395); Switzerland (n = 280); Austria (n = 197); Russia (n = 155); Belgium (n = 102); Ireland (n = 59); Portugal (n = 51); Netherlands (n = 44); Finland (n = 31); Sweden (n = 27); Romania (n = 24); Ukraine (n = 21); Belarus (n = 13); Luxembourg (n = 13); Estonia (n = 12); Norway (n = 12); Czechia (n = 11); Bulgaria (n = 9); Hungary (n = 9); Latvia (n = 8); Poland (n = 7); Andorra (n = 6); Bosnia and Herzegovina (n   =  6); Croatia (n = 6); Iceland (n = 5); Serbia (n = 5); Greece (n = 4); Lithuania (n = 3); Slovakia (n = 3); Slovenia (n = 3); Albania (n = 2); Malta (n = 2); Liechtenstein (n = 1); Macedonia (n = 1); Monaco (n = 1); and Montenegro (n = 1).

Latin and South America included 33 countries: Mexico (n = 1157); Brazil (n = 806); Colombia (n = 387); Argentina (n = 217); Chile (n = 154); Peru (n = 119); Ecuador (n = 102); Venezuela (n = 67); Guatemala (n = 61); Costa Rica (n = 47); Dominican Republic (n = 41); El Salvador (n = 26); Uruguay (n = 24); Bolivia (n = 23); Honduras (n = 18); Nicaragua (n = 15); Panama (n = 15); Paraguay (n = 11); Trinidad and Tobago (n = 8); Jamaica (n = 6); Haiti (n = 4); Guyana (n = 3); Antigua and Barbuda (n = 2); Dominica (n = 2); Saint Lucia (n = 2); Aruba (n = 1); Bahamas (n = 1); Barbados (n = 1); Belize (n = 1); Cuba (n = 1); Grenada (n = 1); Saint Vincent and the Grenadines (n = 1); and Sint Maarten (n = 1).

North America included 2 countries: United States of America (n = 1254) and Canada (n = 338).

South and East Asia included 18 countries: Japan (n = 290); China (n = 90); India (n = 45); Philippines (n = 36); Singapore (n = 26); Malaysia (n = 17); Hong Kong (n = 8); Taiwan (n = 7); South Korea (n = 5); Thailand (n = 4); Georgia (n = 3); Sri Lanka (n = 3); Armenia (n = 2); Vietnam (n = 2); Burma (n = 1); East Timor (n = 1); Maldives (n = 1); and Nepal (n = 1).

Oceania included 5 countries: Australia (n = 133); New Zealand (n = 34); Fiji (n = 1); Marshall Islands (n = 1); and Micronesia (n = 1).

Africa included 26 countries: South Africa (n = 29); Morocco (n = 16); Algeria (n = 7); Tunisia (n = 7); Kenya (n = 6); Nigeria (n = 6); Senegal (n = 5); Namibia (n = 4); Cameroon (n = 3); Egypt (n = 3); Ghana (n = 3); Benin (n = 2); Central African Republic (n = 2); Mali (n = 2); Mauritius (n = 2); Botswana (n = 1); Cote d'Ivoire (n = 1); Ethiopia (n = 1); Guinea-Bissau (n = 1); Madagascar (n = 1); Mauritania (n = 1); South Sudan (n = 1); Swaziland (n = 1); Tanzania (n = 1); Uganda (n = 1); and Zimbabwe (n = 1).

Middle East and Central Asia included 19 countries: Indonesia (n = 18); United Arab Emirates (n = 14); Israel (n = 13); Kazakhstan (n = 9); Pakistan (n = 8); Turkey (n = 8); Iran (n = 7); Saudi Arabia (n = 6); Bahrain (n = 3); Lebanon (n = 3); Jordan (n = 2); Kyrgyzstan (n = 2); Palestinian Territories (n = 2); Qatar (n = 2); Turkmenistan (n = 2); Afghanistan (n = 1); Iraq (n = 1); Kuwait (n = 1); and Syria (n = 1).


We investigated how effects of political orientation on partner preferences differed between four different regions. Namely, these were Europe  (n = 7,417), Latin and South America  (n = 3,325), North America  (n = 1,592), and South and East Asia  (n = 542). We could not include Oceania  (n = 170), Africa  (n = 108), and Middle East and Central Asia  (n = 103) because sample sizes were too small (n < 500) to reach any conclusions.

```{r}
regions = regions %>% filter(participants > 500)
data_included_documented_reg = data_included_documented %>%
  filter(region_final %in% regions$region_final)
```

```{r}
regions_reg =
  data_included_documented_reg %>%
  select(region_final, country) %>%
  table() %>%
  as.data.frame() %>%
  filter(Freq != 0) %>%
  arrange(-Freq)


regions_reg
```


### Analyses {.tabset .active}
#### Political, Ethnic, and Religious Similarity {.tabset}
##### H1a Preference for Similarity in Political Beliefs {.tabset}
###### H1a(1) Linear Effect
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_politicalsim = data_included_documented_reg_wide %>%
  select(-pref_politicalsim) %>%
  map(~lm(scale(data_included_documented_reg_wide$pref_politicalsim) ~ scale(.x),
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_politicalsim_lin_coef = models_pref_politicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_politicalsim_lin_se = models_pref_politicalsim %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_politicalsim_lin_analyses = left_join(models_pref_politicalsim_lin_coef,
                                            models_pref_politicalsim_lin_se,
                                            by = "name") %>%
  mutate(outcome = "H2a) Prefered Political Similarity - Linear Effect")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_politicalsim)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_politicalsim_lin_analyses$n = countries_reg$Freq
data_included_documented_reg %>% filter(!is.na(pref_politicalsim)) %>% nrow()

model = mvmeta(mean ~ 1, data = models_pref_politicalsim_lin_analyses, S = se^2,
               method = "fixed")
summary(model)
```

###### H1a(2) Quadratic Effect: Regression 1 (x <= breaking_point)

```{r}
data_included_documented_reg_wide_reg1 = data_included_documented_reg %>%
  dplyr::filter(political_orientation <= 3) %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_politicalsim_reg1 = data_included_documented_reg_wide_reg1 %>%
  select(-pref_politicalsim) %>%
  map(~lm(scale(data_included_documented_reg_wide_reg1$pref_politicalsim) ~
            scale(.x),
          data = data_included_documented_reg_wide_reg1)) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg1 = models_pref_politicalsim_reg1 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)


models_pref_politicalsim_quad_se_reg1 = models_pref_politicalsim_reg1 %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_politicalsim_quad_analyses_reg1 = left_join(models_pref_politicalsim_quad_coef_reg1,
                                            models_pref_politicalsim_quad_se_reg1,
                                            by = "name") %>%
  mutate(outcome = "H2a(1)) Preferred Political Similarity - Quadratic Effect Regression 1")

models_pref_politicalsim_quad_analyses_reg1$n = countries_reg$Freq

data_included_documented_reg %>% filter(political_orientation <= 3, !is.na(pref_politicalsim)) %>% nrow()

model = mvmeta(mean ~ 1, data = models_pref_politicalsim_quad_analyses_reg1, S = se^2,
               method = "fixed")
summary(model)
```

###### H1a(2) Quadratic Effect: Regression 2 (x >= breaking_point)
```{r}
data_included_documented_reg_wide_reg2 = data_included_documented_reg %>%
  dplyr::filter(political_orientation >= 3) %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_politicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_politicalsim_reg2 = data_included_documented_reg_wide_reg2 %>%
  select(-pref_politicalsim) %>%
  map(~lm(scale(data_included_documented_reg_wide_reg2$pref_politicalsim) ~
            scale(.x),
          data = data_included_documented_reg_wide_reg2)) %>%
  map(lm.beta)

models_pref_politicalsim_quad_coef_reg2 = models_pref_politicalsim_reg2 %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname == "scale(.x)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)


models_pref_politicalsim_quad_se_reg2 = models_pref_politicalsim_reg1 %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_politicalsim_quad_analyses_reg2 = left_join(models_pref_politicalsim_quad_coef_reg2,
                                            models_pref_politicalsim_quad_se_reg2,
                                            by = "name") %>%
  mutate(outcome = "H2a(1)) Preferred Political Similarity - Quadratic Effect Regression 2")

models_pref_politicalsim_quad_analyses_reg2$n = countries_reg$Freq

data_included_documented_reg %>% filter(political_orientation >= 3, !is.na(pref_politicalsim)) %>% nrow()

model = mvmeta(mean ~ 1, data = models_pref_politicalsim_quad_analyses_reg2, S = se^2,
               method = "fixed")
summary(model)
```

##### H1b Preference for Similarity in Ethnicity/Race {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_ethnicalsim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_ethnicalsim = data_included_documented_reg_wide %>%
  select(-pref_ethnicalsim) %>%
  map(~lm(data_included_documented_reg_wide$pref_ethnicalsim ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_ethnicalsim_coef = models_pref_ethnicalsim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_ethnicalsim_se = models_pref_ethnicalsim %>%
 map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_ethnicalsim_analyses = left_join(models_pref_ethnicalsim_coef,
                                            models_pref_ethnicalsim_se,
                                            by = "name") %>%
  mutate(outcome = "H2b) Preferred Ethnic Similarity")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_ethnicalsim)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_ethnicalsim_analyses$n = countries_reg$Freq
sum(models_pref_ethnicalsim_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_ethnicalsim_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H1c Preference for Similarity in Religion {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_religioussim, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_religioussim = data_included_documented_reg_wide %>%
  select(-pref_religioussim) %>%
  map(~lm(data_included_documented_reg_wide$pref_religioussim ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_religioussim_coef = models_pref_religioussim %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_religioussim_se = models_pref_religioussim %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_religioussim_analyses = left_join(models_pref_religioussim_coef,
                                            models_pref_religioussim_se,
                                            by = "name") %>%
  mutate(outcome = "H2c) Preferred Religious Similarity")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_religioussim)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_religioussim_analyses$n = countries_reg$Freq
sum(models_pref_religioussim_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_religioussim_analyses, S = se^2,
               method = "fixed")
summary(model)
```

#### Ideal Partner Preferences {.tabset}
##### H2a Preference for the Level of Financial Security- Successfulness {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_level_financially_secure_successful_ambitious, Europe, 'South and East Asia', 'Latin and South America',
         'North America')

models_pref_level_financially_secure_successful_ambitious = data_included_documented_reg_wide %>%
  select(-pref_level_financially_secure_successful_ambitious) %>%
  map(~lm(data_included_documented_reg_wide$pref_level_financially_secure_successful_ambitious ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_level_financially_secure_successful_ambitious_coef = models_pref_level_financially_secure_successful_ambitious %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_financially_secure_successful_ambitious_se = models_pref_level_financially_secure_successful_ambitious %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))

models_pref_level_financially_secure_successful_ambitious_analyses = left_join(models_pref_level_financially_secure_successful_ambitious_coef,
                                            models_pref_level_financially_secure_successful_ambitious_se,
                                            by = "name") %>%
  mutate(outcome = "H3a) Financial Security-Successfulness")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_level_financially_secure_successful_ambitious)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_level_financially_secure_successful_ambitious_analyses$n = countries_reg$Freq
sum(models_pref_level_financially_secure_successful_ambitious_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_level_financially_secure_successful_ambitious_analyses, S = se^2,
               method = "fixed")
summary(model)
```


##### H2b Preference for the Level of Confidence-Assertiveness {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_level_confident_assertive, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_confident_assertive = data_included_documented_reg_wide %>%
  select(-pref_level_confident_assertive) %>%
  map(~lm(data_included_documented_reg_wide$pref_level_confident_assertive ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_level_confident_assertive_coef = models_pref_level_confident_assertive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_confident_assertive_se = models_pref_level_confident_assertive %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_level_confident_assertive_analyses = left_join(models_pref_level_confident_assertive_coef,
                                            models_pref_level_confident_assertive_se,
                                            by = "name") %>%
  mutate(outcome = "H3d) Confidence-Assertiveness")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_level_confident_assertive)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_level_confident_assertive_analyses$n = countries_reg$Freq
sum(models_pref_level_confident_assertive_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_level_confident_assertive_analyses, S = se^2,
               method = "fixed")
summary(model)
```


##### H2c Preference for the Level of Education-Intelligence {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_level_intelligence_educated, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_intelligence_educated = data_included_documented_reg_wide %>%
  select(-pref_level_intelligence_educated) %>%
  map(~lm(data_included_documented_reg_wide$pref_level_intelligence_educated ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_level_intelligence_educated_coef = models_pref_level_intelligence_educated %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_intelligence_educated_se = models_pref_level_intelligence_educated %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_level_intelligence_educated_analyses = left_join(models_pref_level_intelligence_educated_coef,
                                            models_pref_level_intelligence_educated_se,
                                            by = "name") %>%
  mutate(outcome = "H3e) Education-Intelligence")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_level_intelligence_educated)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_level_intelligence_educated_analyses$n = countries_reg$Freq
sum(models_pref_level_intelligence_educated_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_level_intelligence_educated_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H2d Preference for the Level of Kindness-Supportiveness {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_level_kind_supportive, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_kind_supportive = data_included_documented_reg_wide %>%
  select(-pref_level_kind_supportive) %>%
  map(~lm(data_included_documented_reg_wide$pref_level_kind_supportive ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_level_kind_supportive_coef = models_pref_level_kind_supportive %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_kind_supportive_se = models_pref_level_kind_supportive %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_level_kind_supportive_analyses = left_join(models_pref_level_kind_supportive_coef,
                                            models_pref_level_kind_supportive_se,
                                            by = "name") %>%
  mutate(outcome = "H3b) Kindness-Supportiveness")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_level_kind_supportive)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_level_kind_supportive_analyses$n = countries_reg$Freq
sum(models_pref_level_kind_supportive_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_level_kind_supportive_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H2e Preference for the Level of Attractiveness {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(pref_level_attractiveness, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_pref_level_attractiveness = data_included_documented_reg_wide %>%
  select(-pref_level_attractiveness) %>%
  map(~lm(data_included_documented_reg_wide$pref_level_attractiveness ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_pref_level_attractiveness_coef = models_pref_level_attractiveness %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_pref_level_attractiveness_se = models_pref_level_attractiveness %>%
 map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_pref_level_attractiveness_analyses = left_join(models_pref_level_attractiveness_coef,
                                            models_pref_level_attractiveness_se,
                                            by = "name") %>%
  mutate(outcome = "H3c) Attractiveness")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(pref_level_attractiveness)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_pref_level_attractiveness_analyses$n = countries_reg$Freq
sum(models_pref_level_attractiveness_analyses$n)

model = mvmeta(mean ~ 1, data = models_pref_level_attractiveness_analyses, S = se^2,
               method = "fixed")
summary(model)
```

#### Ideal Age and Height {.tabset}
##### H3a(1) Ideal Age (Importance) {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(imp_age, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_imp_age = data_included_documented_reg_wide %>%
  select(-imp_age) %>%
  map(~lm(data_included_documented_reg_wide$imp_age ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_imp_age_coef = models_imp_age %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_age_se = models_imp_age %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_imp_age_analyses = left_join(models_imp_age_coef,
                                            models_imp_age_se,
                                            by = "name") %>%
  mutate(outcome = "H4a(1)) Ideal Age (Importance)")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(imp_age)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_imp_age_analyses$n = countries_reg$Freq
sum(models_imp_age_analyses$n)

model = mvmeta(mean ~ 1, data = models_imp_age_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H3a(2) Ideal Age (Level) {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(ideal_age_rel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_ideal_age_rel = data_included_documented_reg_wide %>%
  select(-ideal_age_rel) %>%
  map(~lm(data_included_documented_reg_wide$ideal_age_rel ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_ideal_age_rel_coef = models_ideal_age_rel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_age_rel_se = models_ideal_age_rel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_ideal_age_rel_analyses = left_join(models_ideal_age_rel_coef,
                                            models_ideal_age_rel_se,
                                            by = "name") %>%
  mutate(outcome = "H4a(2)) Ideal Age (Level)")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(ideal_age_rel)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_ideal_age_rel_analyses$n = countries_reg$Freq
sum(models_ideal_age_rel_analyses$n)

model = mvmeta(mean ~ 1, data = models_ideal_age_rel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H3b(1) Ideal Height (Importance) {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(imp_height, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_imp_height = data_included_documented_reg_wide %>%
  select(-imp_height) %>%
  map(~lm(data_included_documented_reg_wide$imp_height ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_imp_height_coef = models_imp_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_imp_height_se = models_imp_height %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_imp_height_analyses = left_join(models_imp_height_coef,
                                            models_imp_height_se,
                                            by = "name") %>%
  mutate(outcome = "H4b(1)) Ideal Height (Importance)")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(imp_height)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_imp_height_analyses$n = countries_reg$Freq
sum(models_imp_height_analyses$n)

model = mvmeta(mean ~ 1, data = models_imp_height_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### H3b(2) Ideal Height (Level) {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(ideal_height, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_ideal_height = data_included_documented_reg_wide %>%
  select(-ideal_height) %>%
  map(~lm(data_included_documented_reg_wide$ideal_height ~ .x,
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_ideal_height_coef = models_ideal_height %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_ideal_height_se = models_ideal_height %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))


models_ideal_height_analyses = left_join(models_ideal_height_coef,
                                            models_ideal_height_se,
                                            by = "name") %>%
  mutate(outcome = "H4b(2)) Ideal Height (Level)")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(ideal_height)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_ideal_height_analyses$n = countries_reg$Freq
sum(models_ideal_height_analyses$n)

model = mvmeta(mean ~ 1, data = models_ideal_height_analyses, S = se^2,
               method = "fixed")
summary(model)
```
