==> 12_supplement.Rmd <==
---
title: <font color="#66C2A5">Supplementary Analyses</font>
csl: apa-custom-no-issue.csl
output: 
  html_document:
    code_folding: "show"
editor_options: 
  chunk_output_type: console
---

## {.tabset}
Here, we report on our results on Preferred Relationship Options including Interest in Single Life; Interest in Sexual, Non-Romantic Relationships; Interest in Non-Monogamous Relationships; and Interest in Monogamous Relationships 

### Library
```{r Library}
library(formr)
library(effects)
library(effectsize)
library(lme4)
library(sjstats)
library(lmerTest)
library(dplyr)
library(ggplot2)
library(RColorBrewer)
library(tidyverse)
library(mgcv)         
library(stringr)     
library(sandwich)     
library(lmtest) 
library(minqa)
library(tidyr)
library(ggpubr)
library(coefplot)
library(tibble)
library(purrr) # for running multiple regression
library(broom)
library(mvmeta)
library(lm.beta)
library(countrycode)

apatheme = theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.border = element_blank(),
        axis.line = element_line(),
        legend.title = element_blank(),
        plot.title = element_text(hjust = 0.5))
```

### Main Analysis {.tabset .active}
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### S1 Interest in Single Life {.tabset}
S1 There is a negative linear link between right-wing political orientation and women’s interest in being single.  
Outcome: Ratings for interest in single life. Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_interest_single <- lmer(interest_single ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_single)
confint(model_interest_single, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_single, method = "basic", ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_single ~ political_orientation + (1 | country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S2 Interest in Sexual, Non-Romantic Relationships {.tabset}
S2 There is a negative linear link between right-wing political orientation and women’s interest in having sexual, non-romantic relationships (e.g., hookups, one-night stands).
Outcome: Ratings for interest in having sexual, non-romantic relationships. Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_interest_sexrel <- lmer(interest_sexrel ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_sexrel)
confint(model_interest_sexrel, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_sexrel, method = "basic", ci = 0.997)
#ci(standardize_parameters(model_interest_sexrel, method = "basic"), ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_sexrel ~ political_orientation + (1 + political_orientation| country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S3 Interest in Non-Monogamous Relationships {.tabset}
S3 There is a negative linear link between right-wing political orientation and women’s interest in having non-monogamous relationships.
Outcome: Ratings for interest in having non-monogamous relationships. Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_interest_nonmonrel <- lmer(interest_nonmonrel ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_nonmonrel)
confint(model_interest_nonmonrel, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_nonmonrel, method = "basic", ci = 0.997)
#ci(standardize_parameters(model_interest_nonmonrel, method = "basic"), ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_nonmonrel ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S4 Interest in Monogamous Relationships {.tabset}
S4 There is no link between right-wing political orientation and women’s interest in having monogamous relationships.
Outcome: Ratings for interest in having monogamous relationships. Predictor: Political Orientation. Random intercept and random slope for country.

##### Models
```{r}
model_interest_monrel <- lmer(interest_monrel ~ political_orientation + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_monrel)
confint(model_interest_monrel, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_monrel, method = "basic", ci = 0.997)

#ci(standardize_parameters(model_interest_monrel, method = "basic"), ci = 0.997) #calculate by hand to get exact numbers
```

#####  Plot {.active}
```{r}
lmer(interest_monrel ~ political_orientation + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

### Robustness Analysis
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### S1 Interest in Single Life {.tabset .active}
S1 There is a negative linear link between right-wing political orientation and women’s interest in being single.  
Outcome: Ratings for interest in single life. 
Predictors: Political Orientation & Age. Random intercept and random slope for country.

##### Models
```{r}
model_interest_single_robustcheck <- lmer(interest_single ~ political_orientation + age + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_single_robustcheck)
confint(model_interest_single_robustcheck, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_single_robustcheck, method = "basic", ci = 0.997)
#ci(standardize_parameters(model_interest_single_robustcheck, method = "basic"), ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_single ~ political_orientation + age + (1+political_orientation|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S2 Interest in Sexual, Non-Romantic Relationships {.tabset}
S2 There is a negative linear link between right-wing political orientation and women’s interest in having sexual, non-romantic relationships (e.g., hookups, one-night stands).
Outcome: Ratings for interest in having sexual, non-romantic relationships. Predictors: Political Orientation & Age. Random intercept and random slope for country.

##### Models
```{r}
model_interest_sexrel_robustcheck <- lmer(interest_sexrel ~ political_orientation + age + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_sexrel_robustcheck)
confint(model_interest_sexrel_robustcheck, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_sexrel_robustcheck, method = "basic", ci = 0.997)

#ci(standardize_parameters(model_interest_sexrel_robustcheck, method = "basic"), ci = 0.997)
#calculate by hand to get exact numbers
```

##### Plot {.active}
```{r}
lmer(interest_sexrel ~ political_orientation + age + (1|country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S3 Interest in Non-Monogamous Relationships {.tabset}
S3 There is a negative linear link between right-wing political orientation and women’s interest in having non-monogamous relationships.
Outcome: Ratings for interest in having non-monogamous relationships. Predictors: Political Orientation & Age. Random intercept and random slope for country.

##### Models
```{r}
model_interest_nonmonrel_robustcheck <- lmer(interest_nonmonrel ~ political_orientation + age + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_nonmonrel_robustcheck)
confint(model_interest_nonmonrel_robustcheck, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_nonmonrel_robustcheck, method = "basic", ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_nonmonrel ~ political_orientation + age + (1 | country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```

#### S4 Interest in Monogamous Relationships {.tabset}
S4 There is no link between right-wing political orientation and women’s interest in having monogamous relationships.
Outcome: Ratings for interest in having monogamous relationships. 
Predictors: Political Orientation & Age. Random intercept and random slope for country.

##### Models
```{r}
model_interest_monrel_robustcheck <- lmer(interest_monrel ~ political_orientation + age + (1+political_orientation|country), data = data_included_documented, control =lmerControl(optimizer = "bobyqa"))
```

##### Summary
```{r}
summary(model_interest_monrel_robustcheck)
confint(model_interest_monrel_robustcheck, level = 0.997)
```

##### Standardized Coefficients
```{r}
standardize_parameters(model_interest_monrel_robustcheck, method = "basic", ci = 0.997)
```

##### Plot {.active}
```{r}
lmer(interest_monrel ~ political_orientation + age + (1| country),
       data = data_included_documented) %>%
    allEffects() %>%
    plot()
```


### Exploratory Analysis for Countries
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### Inclusion of Data
```{r}
countries = as.data.frame(table(data_included_documented$country)) %>%
  arrange(-Freq)

countries
```
We will include all countries with more than 500 participants. This allows us to show effect sizes for a diverse range of countries.
Diversity of countries is indicated by:

* location: European (France, Germany, Italy, Spain); North American (United States of America); South American (Mexico, Brazil)
* language: French (France); German (Germany); English (United States of America); Spanish (Mexico, Spain); Italian (Italy); Portuguese (Brazil)
* culture: Western (France, Germany, Italy, Spain, United States of America); Non-Western (Mexico, Brazil)

Sample sizes of other countries are too small (n < 500) to reach any conclusions.

```{r}
seven_countries = countries %>% filter(Freq > 500)
data_included_documented_rescon = data_included_documented %>%
  filter(country %in% seven_countries$Var1)
```

```{r}
countries_rescon =
  data_included_documented_rescon %>%
  select(country) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)


countries_rescon
```

#### Models {.tabset}
##### S1 Interest in Single Life
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_single, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_single = data_included_documented_rescon_wide %>%
  select(-interest_single) %>%
  map(~lm(data_included_documented_rescon_wide$interest_single ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_single_coef = models_interest_single %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_single_se = models_interest_single %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "United States of America",
                       "United.States.of.America", name))



models_interest_single_analyses = left_join(models_interest_single_coef,
                                            models_interest_single_se,
                                            by = "name") %>%
  mutate(outcome = "S1 Interest in Single Life")

countries_rescon =
  data_included_documented_rescon %>%
  filter(!is.na(interest_single)) %>%
  select(country) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_interest_single_analyses$n = countries_rescon$Freq
sum(models_interest_single_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_single_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S2 Interest in Sexual, Non-Romantic Relationships 
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_sexrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_sexrel = data_included_documented_rescon_wide %>%
  select(-interest_sexrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_sexrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_sexrel_coef = models_interest_sexrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_sexrel_se = models_interest_sexrel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "United States of America",
                       "United.States.of.America", name))



models_interest_sexrel_analyses = left_join(models_interest_sexrel_coef,
                                            models_interest_sexrel_se,
                                            by = "name") %>%
  mutate(outcome = "S2 Interest in Sexual, Non-Romantic Relationships")

countries_rescon =
  data_included_documented_rescon %>%
  filter(!is.na(interest_sexrel)) %>%
  select(country) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_interest_sexrel_analyses$n = countries_rescon$Freq
sum(models_interest_sexrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_sexrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S3 Interest in Non-Monogamous Relationships {.tabset}
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_nonmonrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_nonmonrel = data_included_documented_rescon_wide %>%
  select(-interest_nonmonrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_nonmonrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_nonmonrel_coef = models_interest_nonmonrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_nonmonrel_se = models_interest_nonmonrel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "United States of America",
                       "United.States.of.America", name))



models_interest_nonmonrel_analyses = left_join(models_interest_nonmonrel_coef,
                                            models_interest_nonmonrel_se,
                                            by = "name") %>%
  mutate(outcome = "S3 Interest in Non-Monogamous Relationships")

countries_rescon =
  data_included_documented_rescon %>%
  filter(!is.na(interest_nonmonrel)) %>%
  select(country) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_interest_nonmonrel_analyses$n = countries_rescon$Freq
sum(models_interest_nonmonrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_nonmonrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S4 Interest in Monogamous Relationships {.tabset}
```{r}
data_included_documented_rescon_wide = data_included_documented_rescon %>%
  pivot_wider(names_from = country, values_from = political_orientation) %>%
  select(interest_monrel, France, Germany, `United States of America`, Mexico,
         Italy, Brazil, Spain)

models_interest_monrel = data_included_documented_rescon_wide %>%
  select(-interest_monrel) %>%
  map(~lm(data_included_documented_rescon_wide$interest_monrel ~ .x,
      data = data_included_documented_rescon_wide)) %>%
  map(lm.beta)

models_interest_monrel_coef = models_interest_monrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_monrel_se = models_interest_monrel %>%
 map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "United States of America",
                       "United.States.of.America", name))



models_interest_monrel_analyses = left_join(models_interest_monrel_coef,
                                            models_interest_monrel_se,
                                            by = "name") %>%
  mutate(outcome = "S4 Interest in Monogamous Relationships")

countries_rescon =
  data_included_documented_rescon %>%
  filter(!is.na(interest_monrel)) %>%
  select(country) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)

models_interest_monrel_analyses$n = countries_rescon$Freq
sum(models_interest_monrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_monrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

### Exploratory Analysis for Regions {.tabset}
#### Data
Load selected data based on 03_codebook
```{r}
data_included_documented = read.csv(file = "data_included_documented.csv")[,-1]
```

#### Inclusion of Data
```{r}
data_included_documented <- data_included_documented %>%
  mutate(region1 = countrycode(country,
                                 origin = "country.name",
                                 destination = "region"),
         region1 = ifelse(country == "Micronesia",
                            "East Asia and Pacific",
                            region1),
         region2 = countrycode(country,
                                 origin = "country.name",
                                 destination = "continent"),
         region2 = ifelse(country == "Micronesia",
                            "Oceania",
                            region2))

table(data_included_documented$region1)
table(data_included_documented$region2)

data_included_documented = data_included_documented %>%
  mutate(region_final = ifelse(region2 == "Africa", "Africa",
                                  ifelse(region2 == "Asia", "Asia",
                                         ifelse(region2 == "Europe", "Europe",
                           ifelse(region2 == "Oceania", "Oceania",
                                  ifelse(region1 == "Latin America & Caribbean",
                                         "Latin and South America",
                                         ifelse(region1 == "North America",
                                                "North America", NA)))))))

data_included_documented <- data_included_documented %>%
  mutate(region_final = ifelse(country %in% c("Afghanistan", "Bahrain", "Iran",
                                                 "Iraq", "Israel", "Jordan", 
                                                 "Kazakhstan", "Kuwait", "Kyrgyzstan",
                                                 "Lebanon", "Pakistan", 
                                                 "Palestinian Territories", "Qatar",
                                                 "Saudi Arabia", "Syria", 
                                                 "Turkey", "Turkmenistan",
                                                 "United Arab Emirates", "Indonesia"),
                                  "Middle East and Central Asia", region_final),
         region_final = ifelse(country %in% c("Armenia", "Burma",
                                                 "China",
                                                 "East Timor (see Timor-Leste)", 
                                                 "Georgia", "Hong Kong", "India",
                                                 "Japan", "Malaysia", "Maldives",
                                                 "Nepal", "Philippines", "Singapore",
                                                 "South Korea", "Sri Lanka", 
                                                 "Taiwan", "Thailand", "Vietnam"), 
                                  "South and East Asia", region_final))

x <- data_included_documented %>% 
  select(country, region_final) %>%
  group_by(country, region_final) %>%
  summarize(n = n()) %>%
  arrange(region_final, country)

write.csv2(x, file = "country_regions.csv")

regions <- x %>%
  group_by(region_final) %>%
  summarize(countries = n(),
            participants = sum(n))

regions
```
We will include all regions with more than 500 participants. This allows us to show effect sizes for a diverse range of regions.

These regions include the following countries (n)

Europe included 41 countries: France (n = 2,013); Germany (n = 1,846); Italy (n = 968); Spain (n = 562); United Kingdom (n = 499); Denmark (n = 395); Switzerland (n = 280); Austria (n = 197); Russia (n = 155); Belgium (n = 102); Ireland (n = 59); Portugal (n = 51); Netherlands (n = 44); Finland (n = 31); Sweden (n = 27); Romania (n = 24); Ukraine (n = 21); Belarus (n = 13); Luxembourg (n = 13); Estonia (n = 12); Norway (n = 12); Czechia (n = 11); Bulgaria (n = 9); Hungary (n = 9); Latvia (n = 8); Poland (n = 7); Andorra (n = 6); Bosnia and Herzegovina (n   =  6); Croatia (n = 6); Iceland (n = 5); Serbia (n = 5); Greece (n = 4); Lithuania (n = 3); Slovakia (n = 3); Slovenia (n = 3); Albania (n = 2); Malta (n = 2); Liechtenstein (n = 1); Macedonia (n = 1); Monaco (n = 1); and Montenegro (n = 1).

Latin and South America included 33 countries: Mexico (n = 1157); Brazil (n = 806); Colombia (n = 387); Argentina (n = 217); Chile (n = 154); Peru (n = 119); Ecuador (n = 102); Venezuela (n = 67); Guatemala (n = 61); Costa Rica (n = 47); Dominican Republic (n = 41); El Salvador (n = 26); Uruguay (n = 24); Bolivia (n = 23); Honduras (n = 18); Nicaragua (n = 15); Panama (n = 15); Paraguay (n = 11); Trinidad and Tobago (n = 8); Jamaica (n = 6); Haiti (n = 4); Guyana (n = 3); Antigua and Barbuda (n = 2); Dominica (n = 2); Saint Lucia (n = 2); Aruba (n = 1); Bahamas (n = 1); Barbados (n = 1); Belize (n = 1); Cuba (n = 1); Grenada (n = 1); Saint Vincent and the Grenadines (n = 1); and Sint Maarten (n = 1).

North America included 2 countries: United States of America (n = 1254) and Canada (n = 338).

South and East Asia included 18 countries: Japan (n = 290); China (n = 90); India (n = 45); Philippines (n = 36); Singapore (n = 26); Malaysia (n = 17); Hong Kong (n = 8); Taiwan (n = 7); South Korea (n = 5); Thailand (n = 4); Georgia (n = 3); Sri Lanka (n = 3); Armenia (n = 2); Vietnam (n = 2); Burma (n = 1); East Timor (n = 1); Maldives (n = 1); and Nepal (n = 1).

Oceania included 5 countries: Australia (n = 133); New Zealand (n = 34); Fiji (n = 1); Marshall Islands (n = 1); and Micronesia (n = 1).

Africa included 26 countries: South Africa (n = 29); Morocco (n = 16); Algeria (n = 7); Tunisia (n = 7); Kenya (n = 6); Nigeria (n = 6); Senegal (n = 5); Namibia (n = 4); Cameroon (n = 3); Egypt (n = 3); Ghana (n = 3); Benin (n = 2); Central African Republic (n = 2); Mali (n = 2); Mauritius (n = 2); Botswana (n = 1); Cote d'Ivoire (n = 1); Ethiopia (n = 1); Guinea-Bissau (n = 1); Madagascar (n = 1); Mauritania (n = 1); South Sudan (n = 1); Swaziland (n = 1); Tanzania (n = 1); Uganda (n = 1); and Zimbabwe (n = 1).

Middle East and Central Asia included 19 countries: Indonesia (n = 18); United Arab Emirates (n = 14); Israel (n = 13); Kazakhstan (n = 9); Pakistan (n = 8); Turkey (n = 8); Iran (n = 7); Saudi Arabia (n = 6); Bahrain (n = 3); Lebanon (n = 3); Jordan (n = 2); Kyrgyzstan (n = 2); Palestinian Territories (n = 2); Qatar (n = 2); Turkmenistan (n = 2); Afghanistan (n = 1); Iraq (n = 1); Kuwait (n = 1); and Syria (n = 1).


We investigated how effects of political orientation on partner preferences differed between four different regions. Namely, these were Europe  (n = 7,417), Latin and South America  (n = 3,325), North America  (n = 1,592), and South and East Asia  (n = 542). We could not include Oceania  (n = 170), Africa  (n = 108), and Middle East and Central Asia  (n = 103) because sample sizes were too small (n < 500) to reach any conclusions.



```{r}
regions = regions %>% filter(participants > 500)
data_included_documented_reg = data_included_documented %>%
  filter(region_final %in% regions$region_final)
```

```{r}
regions_reg =
  data_included_documented_reg %>%
  select(region_final, country) %>%
  table() %>%
  as.data.frame() %>%
  filter(Freq != 0) %>%
  arrange(-Freq)


regions_reg
```


#### Models {.tabset}
##### S1 Interest in Single Life
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(interest_single, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_single = data_included_documented_reg_wide %>%
  select(-interest_single) %>%
  map(~lm(scale(data_included_documented_reg_wide$interest_single) ~ scale(.x),
      data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_interest_single_coef = models_interest_single %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_single_se = models_interest_single %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
   mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                       "Latin.and.South.America",
                       ifelse(name == "North America",
                       "North.America",
                       name))))

models_interest_single_analyses = left_join(models_interest_single_coef,
                                            models_interest_single_se,
                                            by = "name") %>%
  mutate(outcome = "S1 Interest in Single Life")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(interest_single)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)


models_interest_single_analyses$n = countries_reg$Freq
sum(models_interest_single_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_single_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S2 Interest in Sexual, Non-Romantic Relationships 
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(interest_sexrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_sexrel = data_included_documented_reg_wide %>%
  select(-interest_sexrel) %>%
  map(~lm(scale(data_included_documented_reg_wide$interest_sexrel) ~ scale(.x),
          data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_interest_sexrel_coef = models_interest_sexrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_sexrel_se = models_interest_sexrel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                              "Latin.and.South.America",
                              ifelse(name == "North America",
                                     "North.America",
                                     name))))

models_interest_sexrel_analyses = left_join(models_interest_sexrel_coef,
                                            models_interest_sexrel_se,
                                            by = "name") %>%
  mutate(outcome = "S1 Interest in Single Life")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(interest_sexrel)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)


models_interest_sexrel_analyses$n = countries_reg$Freq
sum(models_interest_sexrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_sexrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S3 Interest in Non-Monogamous Relationships {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(interest_nonmonrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_nonmonrel = data_included_documented_reg_wide %>%
  select(-interest_nonmonrel) %>%
  map(~lm(scale(data_included_documented_reg_wide$interest_nonmonrel) ~ scale(.x),
          data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_interest_nonmonrel_coef = models_interest_nonmonrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_nonmonrel_se = models_interest_nonmonrel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                              "Latin.and.South.America",
                              ifelse(name == "North America",
                                     "North.America",
                                     name))))

models_interest_nonmonrel_analyses = left_join(models_interest_nonmonrel_coef,
                                            models_interest_nonmonrel_se,
                                            by = "name") %>%
  mutate(outcome = "S1 Interest in Single Life")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(interest_nonmonrel)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)


models_interest_nonmonrel_analyses$n = countries_reg$Freq
sum(models_interest_nonmonrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_nonmonrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```

##### S4 Interest in Monogamous Relationships {.tabset}
```{r}
data_included_documented_reg_wide = data_included_documented_reg %>%
  pivot_wider(names_from = region_final, values_from = political_orientation) %>%
  select(interest_monrel, Europe, 'South and East Asia', 'Latin and South America', 'North America')

models_interest_monrel = data_included_documented_reg_wide %>%
  select(-interest_monrel) %>%
  map(~lm(scale(data_included_documented_reg_wide$interest_monrel) ~ scale(.x),
          data = data_included_documented_reg_wide)) %>%
  map(lm.beta)

models_interest_monrel_coef = models_interest_monrel %>%
  map(coef) %>%
  as.data.frame() %>%
  rownames_to_column(var = "rowname") %>%
  filter(rowname != "(Intercept)") %>%
  pivot_longer(cols = -rowname) %>%
  select(-rowname) %>%
  rename(mean = value)

models_interest_monrel_se = models_interest_monrel %>%
  map(tidy) %>%
  tibble(models_pref_politicalsim_lin_se = ., Names = names(.)) %>%
  hoist(models_pref_politicalsim_lin_se, coefficients = "std.error") %>%
  select(-models_pref_politicalsim_lin_se) %>%
  unnest_wider(., coefficients, names_sep = "_") %>%
  select(coefficients_2, Names) %>%
  rename("name" = "Names",
         "se" = "coefficients_2") %>%
  mutate(name = ifelse(name == "South and East Asia",
                       "South.and.East.Asia",
                       ifelse(name == "Latin and South America",
                              "Latin.and.South.America",
                              ifelse(name == "North America",
                                     "North.America",
                                     name))))

models_interest_monrel_analyses = left_join(models_interest_monrel_coef,
                                            models_interest_monrel_se,
                                            by = "name") %>%
  mutate(outcome = "S1 Interest in Single Life")

countries_reg =
  data_included_documented_reg %>%
  filter(!is.na(interest_monrel)) %>%
  select(region_final) %>%
  table() %>%
  as.data.frame() %>%
  arrange(-Freq)


models_interest_monrel_analyses$n = countries_reg$Freq
sum(models_interest_monrel_analyses$n)

model = mvmeta(mean ~ 1, data = models_interest_monrel_analyses, S = se^2,
               method = "fixed")
summary(model)
```
